+++ scratch_test/expt/2D_tests/time_dep/ice_shelf.py
@@ -17,6 +17,8 @@ from grid import *
 sys.path.insert(1, "../../../../solvers/")
 from residuals import make_fo_upwind_residual_function
 from linear_solvers import create_sparse_petsc_la_solver_with_custom_vjp
+from nonlinear_solvers import make_newton_coupled_solver_function,\
+        make_newton_velocity_solver_function_custom_vjp_dynamic_thk
 
                     
 from plotting_stuff import show_vel_field, make_gif, show_damage_field,\
@@ -52,205 +54,316 @@ np.set_printoptions(precision=1, suppress=True, linewidth=np.inf, threshold=np.i
 
 
 
-def make_implicit_advect_scalar_field_fou_residuals_function(nx, ny, dx, dy, vel_bcs="rflc"):
-    if vel_bcs!="rflc":
-        raise "Add the 2 lines of code to enable periodic bcs you lazy bastard, Trys."
-
-    cc_to_fc = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    add_reflection_ghost_cells, add_cont_ghost_cells = add_ghost_cells_fcts(ny, nx)
-
-    def implicit_advect_scalar_field_fou_residuals(u, v, scalar_field,
-                                                   scalar_field_current,
-                                                   source, delta_t, thickness):
-
-        u = linear_extrapolate_over_cf_dynamic_thickness(u, thickness)
-        v = linear_extrapolate_over_cf_dynamic_thickness(v, thickness)
-        h = linear_extrapolate_over_cf_dynamic_thickness(scalar_field_current, thickness)
-        
-        u, v = add_reflection_ghost_cells(u,v)
-        h = add_cont_ghost_cells(h)
-     
-        u_fc_ew, _ = cc_to_fc(u)
-        _, v_fc_ns = cc_to_fc(v)
-
-        u_signs = jnp.where(u_fc_ew>0, 1, -1)
-        v_signs = jnp.where(v_fc_ns>0, 1, -1)
-
-        h_fc_fou_ew = jnp.where(u_fc_ew>0, h[1:-1,:-1], h[1:-1, 1:])
-        h_fc_fou_ns = jnp.where(v_fc_ns>0, h[1:, 1:-1], h[-1:,1:-1])
-
-        ew_flux = u_fc_ew*h_fc_fou_ew
-        ns_flux = v_fc_ns*h_fc_fou_ns
-
-        #remove those ghost cells again!
-        h = h[1:-1,1:-1]
-
-        return ((scalar_field - h)/delta_t - source)*dx*dy +\
-                (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])*dy +\
-                (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])*dx
-
-    return implicit_advect_scalar_field_fou_residuals
-
-
-def make_advect_scalar_field_fou_function(nx, ny, dx, dy, vel_bcs="rflc"):
-    if vel_bcs!="rflc":
-        raise "Add the 2 lines of code to enable periodic bcs you lazy bastard, Trys."
-
-    cc_to_fc = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    add_reflection_ghost_cells, add_cont_ghost_cells = add_ghost_cells_fcts(ny, nx)
-
-    def advect_scalar_field_fou(u, v, scalar_field, source, delta_t, thickness):
-
-        u = linear_extrapolate_over_cf_dynamic_thickness(u, thickness)
-        v = linear_extrapolate_over_cf_dynamic_thickness(v, thickness)
-        h = linear_extrapolate_over_cf_dynamic_thickness(scalar_field, thickness)
-        
-        u, v = add_reflection_ghost_cells(u,v)
-        h = add_cont_ghost_cells(h)
-     
-        u_fc_ew, _ = cc_to_fc(u)
-        _, v_fc_ns = cc_to_fc(v)
-
-        u_signs = jnp.where(u_fc_ew>0, 1, -1)
-        v_signs = jnp.where(v_fc_ns>0, 1, -1)
-
-        h_fc_fou_ew = jnp.where(u_fc_ew>0, h[1:-1,:-1], h[1:-1, 1:])
-        h_fc_fou_ns = jnp.where(v_fc_ns>0, h[1:, 1:-1], h[-1:,1:-1])
-
-        ew_flux = u_fc_ew*h_fc_fou_ew
-        ns_flux = v_fc_ns*h_fc_fou_ns
-
-        #remove those ghost cells again!
-        h = h[1:-1,1:-1]
-
-        h_new = h + delta_t * (
-                    source - \
-                    (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])/dx -\
-                    (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])/dy
-                )
-
-        return jnp.where(thickness>1e-2, h_new, 0)
-    return advect_scalar_field_fou
-
-
-
-def compute_u_v_residuals_function(ny, nx, dy, dx, \
-                                   beta,\
-                                   interp_cc_to_fc,\
-                                   ew_gradient,\
-                                   ns_gradient,\
-                                   cc_gradient,\
-                                   add_uv_ghost_cells,
-                                   add_s_ghost_cells,
-                                   extrp_over_cf):
-    
-    def compute_u_v_residuals(u_1d, v_1d, q, h_1d):
-
-        mucoef = mucoef_0*jnp.exp(q)
-
-        u = u_1d.reshape((ny, nx))
-        v = v_1d.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-
-
-        s_gnd = h + b #b is globally defined
-        s_flt = h * (1-c.RHO_I/c.RHO_W)
-        s = jnp.maximum(s_gnd, s_flt)
-        
-        s = add_s_ghost_cells(s)
-        #jax.debug.print("s: {x}",x=s)
-
-        dsdx, dsdy = cc_gradient(s)
-        #jax.debug.print("dsdx: {x}",x=dsdx)
-        #sneakily fudge this:
-        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
-        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
-        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
-        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
-        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
-        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
-        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
-        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
-
-        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
-        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
-
-
-        #obvs not going to do anything in the no-cf case
-        u = extrp_over_cf(u, h)
-        v = extrp_over_cf(v, h)
-        #momentum_term
-        u, v = add_uv_ghost_cells(u, v)
-
-        #various face-centred derivatives
-        dudx_ew, dudy_ew = ew_gradient(u)
-        dvdx_ew, dvdy_ew = ew_gradient(v)
-        dudx_ns, dudy_ns = ns_gradient(u)
-        dvdx_ns, dvdy_ns = ns_gradient(v)
-
-
-        #interpolate things onto face-cenres
-        h = add_s_ghost_cells(h)
-        h_ew, h_ns = interp_cc_to_fc(h)
-        #remove those ghost cells again!
-        h = h[1:-1,1:-1]
-        #jax.debug.print("h_ew = {x}",x=h_ew)
-        #jax.debug.print("h = {x}",x=h)
-
-        mucoef = add_s_ghost_cells(mucoef)
-        mucoef_ew, mucoef_ns = interp_cc_to_fc(mucoef)
-        #jax.debug.print("mucoef_ew = {x}",x=mucoef_ew)
-
-        #calculate face-centred viscosity:
-        mu_ew = B * mucoef_ew * (dudx_ew**2 + dvdy_ew**2 + dudx_ew*dvdy_ew +\
-                    0.25*(dudy_ew+dvdx_ew)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        mu_ns = B * mucoef_ns * (dudx_ns**2 + dvdy_ns**2 + dudx_ns*dvdy_ns +\
-                    0.25*(dudy_ns+dvdx_ns)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        
-        #to account for calving front boundary condition, set effective viscosities
-        #of faces of all cells with zero thickness to zero:
-        #Again, shouldn't do owt when there's no calving front
-        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h==0, 0, mu_ew[:, 1:]))
-        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h==0, 0, mu_ew[:,:-1]))
-        mu_ns = mu_ns.at[1:, :].set(jnp.where(h==0, 0, mu_ns[1:, :]))
-        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h==0, 0, mu_ns[:-1,:]))
-        
-
-        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
-
-        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
-
-
-        x_mom_residual = visc_x + volume_x
-        y_mom_residual = visc_y + volume_y
-
-        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1)
-
-    return jax.jit(compute_u_v_residuals)
-
-
-#def compute_uvh_residuals_function_lr(ny, nx, dy, dx,
-#                                      beta,
-#                                      interp_cc_to_fc,
-#                                      ew_gradient,
-#                                      ns_gradient,
-#                                      cc_gradient,
-#                                      add_uv_ghost_cells,
-#                                      add_s_ghost_cells):
+#def make_implicit_advect_scalar_field_fou_residuals_function(nx, ny, dx, dy, vel_bcs="rflc"):
+#    if vel_bcs!="rflc":
+#        raise "Add the 2 lines of code to enable periodic bcs you lazy bastard, Trys."
+#
+#    cc_to_fc = interp_cc_with_ghosts_to_fc_function(ny, nx)
+#    add_reflection_ghost_cells, add_cont_ghost_cells = add_ghost_cells_fcts(ny, nx)
+#
+#    def implicit_advect_scalar_field_fou_residuals(u, v, scalar_field,
+#                                                   scalar_field_current,
+#                                                   source, delta_t, thickness):
+#
+#        u = linear_extrapolate_over_cf_dynamic_thickness(u, thickness)
+#        v = linear_extrapolate_over_cf_dynamic_thickness(v, thickness)
+#        h = linear_extrapolate_over_cf_dynamic_thickness(scalar_field_current, thickness)
+#        
+#        u, v = add_reflection_ghost_cells(u,v)
+#        h = add_cont_ghost_cells(h)
+#     
+#        u_fc_ew, _ = cc_to_fc(u)
+#        _, v_fc_ns = cc_to_fc(v)
+#
+#        u_signs = jnp.where(u_fc_ew>0, 1, -1)
+#        v_signs = jnp.where(v_fc_ns>0, 1, -1)
+#
+#        h_fc_fou_ew = jnp.where(u_fc_ew>0, h[1:-1,:-1], h[1:-1, 1:])
+#        h_fc_fou_ns = jnp.where(v_fc_ns>0, h[1:, 1:-1], h[-1:,1:-1])
+#
+#        ew_flux = u_fc_ew*h_fc_fou_ew
+#        ns_flux = v_fc_ns*h_fc_fou_ns
+#
+#        #remove those ghost cells again!
+#        h = h[1:-1,1:-1]
+#
+#        return ((scalar_field - h)/delta_t - source)*dx*dy +\
+#                (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])*dy +\
+#                (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])*dx
+#
+#    return implicit_advect_scalar_field_fou_residuals
+#
+#
+#def make_advect_scalar_field_fou_function(nx, ny, dx, dy, vel_bcs="rflc"):
+#    if vel_bcs!="rflc":
+#        raise "Add the 2 lines of code to enable periodic bcs you lazy bastard, Trys."
+#
+#    cc_to_fc = interp_cc_with_ghosts_to_fc_function(ny, nx)
+#    add_reflection_ghost_cells, add_cont_ghost_cells = add_ghost_cells_fcts(ny, nx)
+#
+#    def advect_scalar_field_fou(u, v, scalar_field, source, delta_t, thickness):
+#
+#        u = linear_extrapolate_over_cf_dynamic_thickness(u, thickness)
+#        v = linear_extrapolate_over_cf_dynamic_thickness(v, thickness)
+#        h = linear_extrapolate_over_cf_dynamic_thickness(scalar_field, thickness)
+#        
+#        u, v = add_reflection_ghost_cells(u,v)
+#        h = add_cont_ghost_cells(h)
+#     
+#        u_fc_ew, _ = cc_to_fc(u)
+#        _, v_fc_ns = cc_to_fc(v)
+#
+#        u_signs = jnp.where(u_fc_ew>0, 1, -1)
+#        v_signs = jnp.where(v_fc_ns>0, 1, -1)
+#
+#        h_fc_fou_ew = jnp.where(u_fc_ew>0, h[1:-1,:-1], h[1:-1, 1:])
+#        h_fc_fou_ns = jnp.where(v_fc_ns>0, h[1:, 1:-1], h[-1:,1:-1])
+#
+#        ew_flux = u_fc_ew*h_fc_fou_ew
+#        ns_flux = v_fc_ns*h_fc_fou_ns
+#
+#        #remove those ghost cells again!
+#        h = h[1:-1,1:-1]
+#
+#        h_new = h + delta_t * (
+#                    source - \
+#                    (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])/dx -\
+#                    (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])/dy
+#                )
+#
+#        return jnp.where(thickness>1e-2, h_new, 0)
+#    return advect_scalar_field_fou
+#
+#
+#
+#def compute_u_v_residuals_function(ny, nx, dy, dx, \
+#                                   beta,\
+#                                   interp_cc_to_fc,\
+#                                   ew_gradient,\
+#                                   ns_gradient,\
+#                                   cc_gradient,\
+#                                   add_uv_ghost_cells,
+#                                   add_s_ghost_cells,
+#                                   extrp_over_cf):
+#    
+#    def compute_u_v_residuals(u_1d, v_1d, q, h_1d):
+#
+#        mucoef = mucoef_0*jnp.exp(q)
+#
+#        u = u_1d.reshape((ny, nx))
+#        v = v_1d.reshape((ny, nx))
+#        h = h_1d.reshape((ny, nx))
+#
+#
+#        s_gnd = h + b #b is globally defined
+#        s_flt = h * (1-c.RHO_I/c.RHO_W)
+#        s = jnp.maximum(s_gnd, s_flt)
+#        
+#        s = add_s_ghost_cells(s)
+#        #jax.debug.print("s: {x}",x=s)
+#
+#        dsdx, dsdy = cc_gradient(s)
+#        #jax.debug.print("dsdx: {x}",x=dsdx)
+#        #sneakily fudge this:
+#        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
+#        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
+#        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
+#        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
+#        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
+#        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
+#        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
+#        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
+#
+#        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
+#        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
+#
+#
+#        #obvs not going to do anything in the no-cf case
+#        u = extrp_over_cf(u, h)
+#        v = extrp_over_cf(v, h)
+#        #momentum_term
+#        u, v = add_uv_ghost_cells(u, v)
+#
+#        #various face-centred derivatives
+#        dudx_ew, dudy_ew = ew_gradient(u)
+#        dvdx_ew, dvdy_ew = ew_gradient(v)
+#        dudx_ns, dudy_ns = ns_gradient(u)
+#        dvdx_ns, dvdy_ns = ns_gradient(v)
+#
+#
+#        #interpolate things onto face-cenres
+#        h = add_s_ghost_cells(h)
+#        h_ew, h_ns = interp_cc_to_fc(h)
+#        #remove those ghost cells again!
+#        h = h[1:-1,1:-1]
+#        #jax.debug.print("h_ew = {x}",x=h_ew)
+#        #jax.debug.print("h = {x}",x=h)
+#
+#        mucoef = add_s_ghost_cells(mucoef)
+#        mucoef_ew, mucoef_ns = interp_cc_to_fc(mucoef)
+#        #jax.debug.print("mucoef_ew = {x}",x=mucoef_ew)
+#
+#        #calculate face-centred viscosity:
+#        mu_ew = B * mucoef_ew * (dudx_ew**2 + dvdy_ew**2 + dudx_ew*dvdy_ew +\
+#                    0.25*(dudy_ew+dvdx_ew)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
+#        mu_ns = B * mucoef_ns * (dudx_ns**2 + dvdy_ns**2 + dudx_ns*dvdy_ns +\
+#                    0.25*(dudy_ns+dvdx_ns)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
+#        
+#        #to account for calving front boundary condition, set effective viscosities
+#        #of faces of all cells with zero thickness to zero:
+#        #Again, shouldn't do owt when there's no calving front
+#        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h==0, 0, mu_ew[:, 1:]))
+#        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h==0, 0, mu_ew[:,:-1]))
+#        mu_ns = mu_ns.at[1:, :].set(jnp.where(h==0, 0, mu_ns[1:, :]))
+#        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h==0, 0, mu_ns[:-1,:]))
+#        
+#
+#        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
+#                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
+#                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
+#                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
+#
+#        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
+#                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
+#                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
+#                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
+#
+#
+#        x_mom_residual = visc_x + volume_x
+#        y_mom_residual = visc_y + volume_y
+#
+#        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1)
+#
+#    return jax.jit(compute_u_v_residuals)
+#
+#
+##def compute_uvh_residuals_function_lr(ny, nx, dy, dx,
+##                                      beta,
+##                                      interp_cc_to_fc,
+##                                      ew_gradient,
+##                                      ns_gradient,
+##                                      cc_gradient,
+##                                      add_uv_ghost_cells,
+##                                      add_s_ghost_cells):
+##    
+##
+##    def compute_uvh_residuals(u, v, h, viscosity, h_t, source, delta_t):
+##        
+##        u = u.reshape((ny, nx))
+##        v = v.reshape((ny, nx))
+##        h = h.reshape((ny, nx))
+##        h_t = h_t.reshape((ny, nx))
+##
+##
+##        ######### MOMENTUM RESIDUALS #############################
+##
+##        s_gnd = h + b #b is globally defined
+##        s_flt = h * (1-c.RHO_I/c.RHO_W)
+##        s = jnp.maximum(s_gnd, s_flt)
+##        
+##        s = add_s_ghost_cells(s)
+##        #jax.debug.print("s: {x}",x=s)
+##
+##        dsdx, dsdy = cc_gradient(s)
+##        #jax.debug.print("dsdx: {x}",x=dsdx)
+##        #sneakily fudge this:
+##        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
+##        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
+##        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
+##        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
+##        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
+##        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
+##        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
+##        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
+##
+##        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
+##        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
+##
+##        mu_ew, mu_ns = viscosity
+##
+##        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
+##                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
+##                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
+##                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
+##
+##        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
+##                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
+##                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
+##                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
+##
+##        x_mom_residual = visc_x + volume_x
+##        y_mom_residual = visc_y + volume_y
+##
+##
+##        #################################################################
+##
+##
+##
+##
+##
+##
+##
+##        ################ ADVECTION RESIDUALS #############################
+##
+##
+##        h = linear_extrapolate_over_cf_dynamic_thickness(h, h)
+##        u = linear_extrapolate_over_cf_dynamic_thickness(u, h)
+##        v = linear_extrapolate_over_cf_dynamic_thickness(v, h)
+##        
+##        
+##
+##        u_fc_ew, _ = interp_cc_to_fc(u)
+##        _, v_fc_ns = interp_cc_to_fc(v)
+##
+##        u_signs = jnp.where(u_fc_ew>0, 1, -1)
+##        v_signs = jnp.where(v_fc_ns>0, 1, -1)
+##
+##
+##        #face-centred values according to first-order upwinding
+##        h_fc_fou_ew = jnp.where(u_fc_ew>0, h[1:-1,:-1], h[1:-1, 1:])
+##        h_fc_fou_ns = jnp.where(v_fc_ns>0, h[1:, 1:-1], h[-1:,1:-1])
+##
+##        ew_flux = u_fc_ew*h_fc_fou_ew
+##        ns_flux = v_fc_ns*h_fc_fou_ns
+##
+##        #remove those ghost cells again!
+##        h = h[1:-1,1:-1]
+##
+##        adv_residual =  ((h - h_t)/delta_t - source)*dx*dy +\
+##                (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])*dy +\
+##                (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])*dx
+##
+##        #################################################################
+##
+##
+##
+##        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1), adv_residual.reshape(-1) 
+##
+##    return compute_uvh_residuals
+#
+#
+#def compute_uvh_residuals_function(ny, nx, dy, dx,
+#                                   beta,
+#                                   interp_cc_to_fc,
+#                                   ew_gradient,
+#                                   ns_gradient,
+#                                   cc_gradient,
+#                                   add_uv_ghost_cells,
+#                                   add_s_ghost_cells):
 #    
 #
-#    def compute_uvh_residuals(u, v, h, viscosity, h_t, source, delta_t):
+#    def compute_uvh_residuals(u_1d, v_1d, h_1d, q, h_t, source, delta_t):
+#
+#        h_static = add_s_ghost_cells(h_t)
+#
+#        #print(h_static)
 #        
-#        u = u.reshape((ny, nx))
-#        v = v.reshape((ny, nx))
-#        h = h.reshape((ny, nx))
-#        h_t = h_t.reshape((ny, nx))
+#        mucoef = mucoef_0*jnp.exp(q)
+#
+#        u = u_1d.reshape((ny, nx))
+#        v = v_1d.reshape((ny, nx))
+#        h = h_1d.reshape((ny, nx))
+#        #h_t = h_t.reshape((ny, nx))
 #
 #
 #        ######### MOMENTUM RESIDUALS #############################
@@ -277,7 +390,41 @@ def compute_u_v_residuals_function(ny, nx, dy, dx, \
 #        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
 #        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
 #
-#        mu_ew, mu_ns = viscosity
+#        #momentum_term
+#        u_ghost, v_ghost = add_uv_ghost_cells(u, v)
+#        u_full = linear_extrapolate_over_cf_dynamic_thickness(u_ghost, h_static)
+#        v_full = linear_extrapolate_over_cf_dynamic_thickness(v_ghost, h_static)
+#
+#        #various face-centred derivatives
+#        dudx_ew, dudy_ew = ew_gradient(u_full)
+#        dvdx_ew, dvdy_ew = ew_gradient(v_full)
+#        dudx_ns, dudy_ns = ns_gradient(u_full)
+#        dvdx_ns, dvdy_ns = ns_gradient(v_full)
+#
+#        #jax.debug.print("dudxew = {x}",x=dudx_ew)
+#
+#        #interpolate things onto face-centres
+#        h_ghost = add_s_ghost_cells(h)
+#        h_ew, h_ns = interp_cc_to_fc(h_ghost)
+#        
+#
+#        mucoef = add_s_ghost_cells(mucoef)
+#        mucoef_ew, mucoef_ns = interp_cc_to_fc(mucoef)
+#        #jax.debug.print("mucoef_ew = {x}",x=mucoef_ew)
+#
+#        #calculate face-centred viscosity:
+#        mu_ew = B * mucoef_ew * (dudx_ew**2 + dvdy_ew**2 + dudx_ew*dvdy_ew +\
+#                    0.25*(dudy_ew+dvdx_ew)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
+#        mu_ns = B * mucoef_ns * (dudx_ns**2 + dvdy_ns**2 + dudx_ns*dvdy_ns +\
+#                    0.25*(dudy_ns+dvdx_ns)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
+#        
+#        #to account for calving front boundary condition, set effective viscosities
+#        #of faces of all cells with zero thickness to zero:
+#        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h_t<1e-2, 0, mu_ew[:, 1:]))
+#        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h_t<1e-2, 0, mu_ew[:,:-1]))
+#        mu_ns = mu_ns.at[1:, :].set(jnp.where(h_t<1e-2, 0, mu_ns[1:, :]))
+#        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h_t<1e-2, 0, mu_ns[:-1,:]))
+#        
 #
 #        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
 #                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
@@ -293,6 +440,10 @@ def compute_u_v_residuals_function(ny, nx, dy, dx, \
 #        y_mom_residual = visc_y + volume_y
 #
 #
+#        #print(f"xmom {jnp.max(jnp.abs(x_mom_residual))}")
+#        #print(f"ymom {jnp.max(jnp.abs(y_mom_residual))}")
+#
+#
 #        #################################################################
 #
 #
@@ -304,724 +455,567 @@ def compute_u_v_residuals_function(ny, nx, dy, dx, \
 #        ################ ADVECTION RESIDUALS #############################
 #
 #
-#        h = linear_extrapolate_over_cf_dynamic_thickness(h, h)
-#        u = linear_extrapolate_over_cf_dynamic_thickness(u, h)
-#        v = linear_extrapolate_over_cf_dynamic_thickness(v, h)
-#        
-#        
+#        hh = linear_extrapolate_over_cf_dynamic_thickness(h_ghost, h_static)
+#        #hh = h.copy()
 #
-#        u_fc_ew, _ = interp_cc_to_fc(u)
-#        _, v_fc_ns = interp_cc_to_fc(v)
+#        #print("-----------------------------")
+#        #print("-----------------------------")
+#        #print("-----------------------------")
+#        #print("-----------------------------")
+#        #print("-----------------------------")
+#        #print("-----------------------------")
+#        #print("-----------------------------")
+#        #print("hstatic")
+#        #print(h_static)
+#        #print("h")
+#        #print(h)
+#        #print("hghost")
+#        #print(h_ghost)
+#        #print("hh")
+#        #print(hh)
+#
+#        u_fc_ew, _ = interp_cc_to_fc(u_full)
+#        _, v_fc_ns = interp_cc_to_fc(v_full)
 #
 #        u_signs = jnp.where(u_fc_ew>0, 1, -1)
 #        v_signs = jnp.where(v_fc_ns>0, 1, -1)
 #
 #
-#        #face-centred values according to first-order upwinding
-#        h_fc_fou_ew = jnp.where(u_fc_ew>0, h[1:-1,:-1], h[1:-1, 1:])
-#        h_fc_fou_ns = jnp.where(v_fc_ns>0, h[1:, 1:-1], h[-1:,1:-1])
-#
-#        ew_flux = u_fc_ew*h_fc_fou_ew
-#        ns_flux = v_fc_ns*h_fc_fou_ns
-#
-#        #remove those ghost cells again!
-#        h = h[1:-1,1:-1]
-#
-#        adv_residual =  ((h - h_t)/delta_t - source)*dx*dy +\
-#                (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])*dy +\
-#                (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])*dx
-#
-#        #################################################################
+#        ##face-centred values according to first-order upwinding
+#        h_fc_fou_ew = jnp.where(u_fc_ew>0, hh[1:-1,:-1], hh[1:-1, 1:])
+#        h_fc_fou_ns = jnp.where(v_fc_ns>0, hh[1:, 1:-1], hh[-1:,1:-1])
 #
 #
 #
-#        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1), adv_residual.reshape(-1) 
+#        flux_term = (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])*dy*delta_t +\
+#                    (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])*dx*delta_t
+#        #to keep calving front in same location, prevent any flux into or out of ice-free cells!
+#        flux_term = jnp.where(h_t>1e-2, flux_term, 0)
+#
+#
+#        adv_residual =  ((h - h_t) - source * delta_t)*dx*dy + flux_term
+#
+#        ##################################################################
+#
+#
+#
+#        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1), adv_residual.reshape(-1)
 #
 #    return compute_uvh_residuals
+#
+#
+#def print_residual_things(residual, rhs, init_residual, i):
+#    old_residual = residual
+#    #residual = jnp.max(jnp.abs(-rhs))
+#    residual = jnp.sqrt(jnp.sum(rhs**2))
+#
+#    if i==0:
+#        init_residual = residual.copy()
+#        print("Initial residual: {}".format(residual))
+#    else:
+#        print("residual: {}".format(residual))
+#        print("Residual reduction factor: {}".format(old_residual/residual))
+#    print("------")
+#
+#    return old_residual, residual, init_residual
+#
+#
+#def make_newton_coupled_solver_function(ny, nx, dy, dx, C, n_iterations):
+#
+#    beta_eff = C.copy()
+#
+#    #functions for various things:
+#    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
+#    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
+#    cc_gradient                                = cc_gradient_function(dy, dx)
+#    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
+#    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
+#    get_uvh_residuals = compute_uvh_residuals_function(ny, nx, dy, dx,\
+#                                                       beta_eff,\
+#                                                       interp_cc_to_fc,\
+#                                                       ew_gradient, ns_gradient,\
+#                                                       cc_gradient,\
+#                                                       add_uv_ghost_cells,\
+#                                                       add_cont_ghost_cells)
+#
+#    #############
+#    #setting up bvs and coords for a single block of the jacobian
+#    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
+#                                                                                  periodic_x=False)
+#
+#    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
+#    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
+#    mask = (i_coordinate_sets>=0)
+#
+#
+#    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
+#                                                        basis_vectors,\
+#                                                        i_coordinate_sets,\
+#                                                        j_coordinate_sets,\
+#                                                        mask,\
+#                                                        3,
+#                                                        active_indices=(0,1,2)
+#                                                       )
+#    #sparse_jacrev = jax.jit(sparse_jacrev)
+#
+#
+#    i_coordinate_sets = i_coordinate_sets[mask]
+#    j_coordinate_sets = j_coordinate_sets[mask]
+#    #############
+#
+#    coords = jnp.stack([
+#        jnp.concatenate(
+#           [i_coordinate_sets,           i_coordinate_sets,           i_coordinate_sets,\
+#            i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),\
+#            i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx)]
+#                       ),\
+#        jnp.concatenate(
+#           [j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
+#            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
+#            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx)]
+#                       )
+#                       ])
+#
+#   
+#    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*3, ny*nx*3),\
+#                                                              ksp_type="gmres",\
+#                                                              preconditioner="hypre",\
+#                                                              precondition_only=False,\
+#                                                              monitor_ksp=False)
+#
+#
+#
+#    comm = PETSc.COMM_WORLD
+#    size = comm.Get_size()
+#    
+#    #@custom_vjp
+#    def solver(q, u_trial, v_trial, h_now, accm, delta_t):
+#        u_trial = jnp.where(h_now>1e-10, u_trial, 0)
+#        v_trial = jnp.where(h_now>1e-10, v_trial, 0)
+#
+#        h_1d = h_now.copy().reshape(-1)
+#        u_1d = u_trial.copy().reshape(-1)
+#        v_1d = v_trial.copy().reshape(-1)
+#
+#        residual = jnp.inf
+#        init_res = 0
+#
+#        old_residual = jnp.inf
+#
+#        #get_uvh_residuals(u_1d, v_1d, h_1d, q, h_now, accm, delta_t)
+#
+#        for i in range(n_iterations):
+#
+#            dRu_du, dRv_du, dRh_du, \
+#            dRu_dv, dRv_dv, dRh_dv, \
+#            dRu_dh, dRv_dh, dRh_dh = sparse_jacrev(get_uvh_residuals, \
+#                                                  (u_1d, v_1d, h_1d, q, h_now, accm, delta_t)
+#                                                          )
+#
+#            nz_jac_values = jnp.concatenate([dRu_du[mask], dRu_dv[mask], dRu_dh[mask],\
+#                                             dRv_du[mask], dRv_dv[mask], dRv_dh[mask],\
+#                                             dRh_du[mask], dRh_dv[mask], dRh_dh[mask]])
+#
+#          
+#            #full_jac = jnp.zeros((ny*nx*3, ny*nx*3))
+#            #full_jac = full_jac.at[coords[0,:], coords[1,:]].set(nz_jac_values)
+#            #print(full_jac)
+#            #raise
+#
+#
+#            rhs = -jnp.concatenate(get_uvh_residuals(u_1d, v_1d, h_1d, q, h_now, accm, delta_t))
+#
+#            #print(jnp.max(rhs))
+#            #raise
+#
+#            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
+#            
+#            
+#
+#
+#            du = la_solver(nz_jac_values, rhs)
+#
+#
+#
+#            #residual = jnp.sqrt(jnp.sum(rhs**2))
+#            #print(residual)
+#            #print(old_residual/residual)
+#            #old_residual = residual.copy()
+#
+#            #print(f"!!! {jnp.max(jnp.abs(du))}")
+#            #
+#
+#            #iptr, j, values = scipy_coo_to_csr(nz_jac_values, coords,\
+#            #                               (ny*nx*3, ny*nx*3), return_decomposition=True)
+#
+#            #A = PETSc.Mat().createAIJ(size=(ny*nx*3, ny*nx*3), \
+#            #                          csr=(iptr.astype(np.int32), j.astype(np.int32), values),\
+#            #                          comm=comm)
+#            #
+#            #b = PETSc.Vec().createWithArray(du, comm=comm)
+#
+#            #jdu = jnp.zeros_like(du)
+#            #pjdu = PETSc.Vec().createWithArray(jdu, comm=comm)
+#            #A.mult(b, pjdu)
+#
+#            #print(f"Jdeltau = {jnp.max(jnp.abs(jdu))}")
+#
+#            #
+#            #print(f"R = {jnp.max(jnp.abs(jdu - rhs))}")
+#
+#
+#
+#
+#
+#
+#
+#            u_1d = u_1d+du[:(ny*nx)]
+#            v_1d = v_1d+du[(ny*nx):(2*ny*nx)]
+#            h_1d = h_1d+du[(2*ny*nx):]
+#
+#
+#
+#
+#        res_final = jnp.max(jnp.abs(jnp.concatenate(
+#                                    get_uvh_residuals(u_1d, v_1d, h_1d,  q, h_now, accm, delta_t)
+#                                                   )
+#                                   )
+#                           )
+#        print("----------")
+#        print("Final residual: {}".format(res_final))
+#        print("Total residual reduction factor: {}".format(init_res/res_final))
+#        print("----------")
+#        
+#        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx)), h_1d.reshape((ny,nx))
+#
+#    return solver
+#
+#
+#
+#
+#def make_quasi_newton_coupled_solver_function(ny, nx, dy, dx, C, n_iterations):
+#
+#    beta_eff = C.copy()
+#
+#    #functions for various things:
+#    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
+#    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
+#    cc_gradient                                = cc_gradient_function(dy, dx)
+#    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
+#    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
+#    get_uvh_residuals = compute_uvh_residuals_function_lr(ny, nx, dy, dx,\
+#                                                          beta_eff,\
+#                                                          interp_cc_to_fc,\
+#                                                          ew_gradient, ns_gradient,\
+#                                                          cc_gradient,\
+#                                                          add_uv_ghost_cells,\
+#                                                          add_cont_ghost_cells)
+#
+#    #############
+#    #setting up bvs and coords for a single block of the jacobian
+#    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
+#                                                                                  periodic_x=False)
+#
+#    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
+#    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
+#    mask = (i_coordinate_sets>=0)
+#
+#
+#    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
+#                                                        basis_vectors,\
+#                                                        i_coordinate_sets,\
+#                                                        j_coordinate_sets,\
+#                                                        mask,\
+#                                                        3,
+#                                                        active_indices=(0,1,2)
+#                                                       )
+#    #sparse_jacrev = jax.jit(sparse_jacrev)
+#
+#
+#    i_coordinate_sets = i_coordinate_sets[mask]
+#    j_coordinate_sets = j_coordinate_sets[mask]
+#    #############
+#
+#    coords = jnp.stack([
+#        jnp.concatenate(
+#           [i_coordinate_sets,           i_coordinate_sets,           i_coordinate_sets,\
+#            i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),\
+#            i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx)]
+#                       ),\
+#        jnp.concatenate(
+#           [j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
+#            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
+#            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx)]
+#                       )
+#                       ])
+#
+#   
+#    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*3, ny*nx*3),\
+#                                                              ksp_type="bcgs",\
+#                                                              preconditioner="hypre",\
+#                                                              precondition_only=False,\
+#                                                              monitor_ksp=False)
+#
+#
+#
+#    #@custom_vjp
+#    def solver(q, u_trial, v_trial, h_now, accm, delta_t):
+#        u_trial = jnp.where(h_now>1e-10, u_trial, 0)
+#        v_trial = jnp.where(h_now>1e-10, v_trial, 0)
+#
+#        h_1d = h_now.copy().reshape(-1)
+#        u_1d = u_trial.copy().reshape(-1)
+#        v_1d = v_trial.copy().reshape(-1)
+#
+#        residual = jnp.inf
+#        init_res = 0
+#
+#
+#        viscosity = fc_viscosity(u_trial, v_trial, q, h_now)
+#
+#
+#        for i in range(n_iterations):
+#
+#            dRu_du, dRv_du, dRh_du, \
+#            dRu_dv, dRv_dv, dRh_dv, \
+#            dRu_dh, dRv_dh, dRh_dh = sparse_jacrev(get_uvh_residuals_lr, \
+#                                        (u_1d, v_1d, h_1d, viscosity, h_now, accm, delta_t)
+#                                                          )
+#
+#            nz_jac_values = jnp.concatenate([dRu_du[mask], dRu_dv[mask], dRu_dh[mask],\
+#                                             dRv_du[mask], dRv_dv[mask], dRv_dh[mask],\
+#                                             dRh_du[mask], dRh_dv[mask], dRh_dh[mask]])
+#
+#          
+#            #full_jac = jnp.zeros((ny*nx*3, ny*nx*3))
+#            #full_jac = full_jac.at[coords[0,:], coords[1,:]].set(nz_jac_values)
+#            #print(full_jac)
+#            #raise
+#
+#
+#            rhs = -jnp.concatenate(get_uvh_residuals_lr(u_1d, v_1d, h_1d,
+#                                                        viscosity, h_now, accm, delta_t
+#                                                        )
+#                                   )
+#
+#            #print(jnp.max(rhs))
+#            #raise
+#
+#            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
+#
+#
+#            du = la_solver(nz_jac_values, rhs)
+#
+#            u_1d = u_1d+du[:(ny*nx)]
+#            v_1d = v_1d+du[(ny*nx):(2*ny*nx)]
+#            h_1d = h_1d+du[(2*ny*nx):]
+#
+#
+#        res_final = jnp.max(jnp.abs(jnp.concatenate(
+#                                get_uvh_residuals_lr(u_1d, v_1d, h_1d, 
+#                                                     viscosity, h_now, accm, delta_t
+#                                                    )
+#                                                   )
+#                                   )
+#                           )
+#
+#        print("----------")
+#        print("Final residual: {}".format(res_final))
+#        print("Total residual reduction factor: {}".format(init_res/res_final))
+#        print("----------")
+#        
+#        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx)), h_1d.reshape((ny,nx))
+#
+#    return solver
+#
+#
+#
+#
+#def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
+#                                                    C, n_iterations):
+#
+#    beta_eff = C.copy()
+#
+#    #functions for various things:
+#    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
+#    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
+#    cc_gradient                                = cc_gradient_function(dy, dx)
+#    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
+#    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
+#    #add_uv_ghost_cells                         = apply_scalar_ghost_cells_to_vector(
+#    #                                                add_ghost_cells_periodic_dirichlet_function(ny,nx)
+#    #                                             )
+#    extrapolate_over_cf                        = linear_extrapolate_over_cf_dynamic_thickness
+#
+#    get_u_v_residuals = compute_u_v_residuals_function(ny, nx, dy, dx,\
+#                                                       beta_eff,\
+#                                                       interp_cc_to_fc,\
+#                                                       ew_gradient, ns_gradient,\
+#                                                       cc_gradient,\
+#                                                       add_uv_ghost_cells,\
+#                                                       add_cont_ghost_cells,\
+#                                                       extrapolate_over_cf)
+#    
+#
+#    #############
+#    #setting up bvs and coords for a single block of the jacobian
+#    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
+#                                                                                  periodic_x=False)
+#
+#    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
+#    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
+#    mask = (i_coordinate_sets>=0)
+#
+#
+#    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
+#                                                        basis_vectors,\
+#                                                        i_coordinate_sets,\
+#                                                        j_coordinate_sets,\
+#                                                        mask,\
+#                                                        2,
+#                                                        active_indices=(0,1)
+#                                                       )
+#    #sparse_jacrev = jax.jit(sparse_jacrev)
+#
+#
+#    i_coordinate_sets = i_coordinate_sets[mask]
+#    j_coordinate_sets = j_coordinate_sets[mask]
+#    #############
+#
+#    coords = jnp.stack([
+#                    jnp.concatenate(
+#                                [i_coordinate_sets,         i_coordinate_sets,\
+#                                 i_coordinate_sets+(ny*nx), i_coordinate_sets+(ny*nx)]
+#                                   ),\
+#                    jnp.concatenate(
+#                                [j_coordinate_sets, j_coordinate_sets+(ny*nx),\
+#                                 j_coordinate_sets, j_coordinate_sets+(ny*nx)]
+#                                   )
+#                       ])
+#
+#   
+#    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*2, ny*nx*2),\
+#                                                              ksp_type="bcgs",\
+#                                                              preconditioner="hypre",\
+#                                                              precondition_only=False)
+#
+#
+#
+#    @custom_vjp
+#    def solver(q, u_trial, v_trial, h):
+#        u_trial = jnp.where(h>1e-10, u_trial, 0)
+#        v_trial = jnp.where(h>1e-10, v_trial, 0)
+#
+#        h_1d = h.copy().reshape(-1)
+#        u_1d = u_trial.copy().reshape(-1)
+#        v_1d = v_trial.copy().reshape(-1)
+#
+#        residual = jnp.inf
+#        init_res = 0
+#
+#        for i in range(n_iterations):
+#
+#            dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
+#                                                           (u_1d, v_1d, q, h_1d)
+#                                                          )
+#
+#            nz_jac_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
+#                                             dJv_du[mask], dJv_dv[mask]])
+#
+#            #full_jac = jnp.zeros((ny*nx*2, ny*nx*2))
+#            #full_jac = full_jac.at[coords[0,:], coords[1,:]].set(nz_jac_values)
+#            #print(full_jac)
+#            #raise
+#
+#
+#            rhs = -jnp.concatenate(get_u_v_residuals(u_1d, v_1d, q, h_1d))
+#
+#            #print(jnp.max(rhs))
+#            #raise
+#
+#            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
+#
+#
+#            du = la_solver(nz_jac_values, rhs)
+#
+#            u_1d = u_1d+du[:(ny*nx)]
+#            v_1d = v_1d+du[(ny*nx):]
+#
+#
+#        res_final = jnp.max(jnp.abs(jnp.concatenate(
+#                                    get_u_v_residuals(u_1d, v_1d, q, h_1d)
+#                                                   )
+#                                   )
+#                           )
+#        print("----------")
+#        print("Final residual: {}".format(res_final))
+#        print("Total residual reduction factor: {}".format(init_res/res_final))
+#        print("----------")
+#        
+#        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx))
+#
+#
+#
+#    def solver_fwd(q, u_trial, v_trial, h):
+#        u_trial = jnp.where(h>1e-10, u_trial, 0)
+#        v_trial = jnp.where(h>1e-10, v_trial, 0)
+#        
+#        u, v = solver(q, u_trial, v_trial, h)
+#
+#        dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
+#                                                       (u.reshape(-1), v.reshape(-1), q, h.reshape(-1))
+#                                                      )
+#        dJ_dvel_nz_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
+#                                             dJv_du[mask], dJv_dv[mask]])
+#
+#
+#        fwd_residuals = (u, v, dJ_dvel_nz_values, q, h)
+#        #fwd_residuals = (u, v, q)
+#
+#        return (u, v), fwd_residuals
+#
+#
+#    def solver_bwd(res, cotangent):
+#        
+#        u, v, dJ_dvel_nz_values, q, h = res
+#        #u, v, q = res
+#
+#        u_bar, v_bar = cotangent
+#        
+#
+#        #dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
+#        #                                               (u.reshape(-1), v.reshape(-1), q)
+#        #                                              )
+#        #dJ_dvel_nz_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
+#        #                                     dJv_du[mask], dJv_dv[mask]])
+#
+#
+#        lambda_ = la_solver(dJ_dvel_nz_values,
+#                            -jnp.concatenate([u_bar, v_bar]),
+#                            transpose=True)
+#
+#        lambda_u = lambda_[:(ny*nx)]
+#        lambda_v = lambda_[(ny*nx):]
+#
+#
+#        #phi_bar = (dG/dphi)^T lambda
+#        _, pullback_function = jax.vjp(get_u_v_residuals,
+#                                         u.reshape(-1), v.reshape(-1), q, h.reshape(-1)
+#                                      )
+#        _, _, mu_bar = pullback_function((lambda_u, lambda_v))
+#        
+##        #bwd has to return a tuple of cotangents for each primal input
+##        #of solver, so have to return this 1-tuple:
+##        return (mu_bar.reshape((ny, nx)), )
+#
+#        #I wonder if I can get away with just returning None for u_trial_bar, v_trial_bar, h_bar...
+#        return (mu_bar.reshape((ny, nx)), None, None, None)
+#
+#
+#    solver.defvjp(solver_fwd, solver_bwd)
+#
+#    return solver
 
-
-def compute_uvh_residuals_function(ny, nx, dy, dx,
-                                   beta,
-                                   interp_cc_to_fc,
-                                   ew_gradient,
-                                   ns_gradient,
-                                   cc_gradient,
-                                   add_uv_ghost_cells,
-                                   add_s_ghost_cells):
-    
-
-    def compute_uvh_residuals(u_1d, v_1d, h_1d, q, h_t, source, delta_t):
-
-        h_static = add_s_ghost_cells(h_t)
-
-        #print(h_static)
-        
-        mucoef = mucoef_0*jnp.exp(q)
-
-        u = u_1d.reshape((ny, nx))
-        v = v_1d.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-        #h_t = h_t.reshape((ny, nx))
-
-
-        ######### MOMENTUM RESIDUALS #############################
-
-        s_gnd = h + b #b is globally defined
-        s_flt = h * (1-c.RHO_I/c.RHO_W)
-        s = jnp.maximum(s_gnd, s_flt)
-        
-        s = add_s_ghost_cells(s)
-        #jax.debug.print("s: {x}",x=s)
-
-        dsdx, dsdy = cc_gradient(s)
-        #jax.debug.print("dsdx: {x}",x=dsdx)
-        #sneakily fudge this:
-        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
-        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
-        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
-        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
-        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
-        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
-        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
-        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
-
-        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
-        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
-
-        #momentum_term
-        u_ghost, v_ghost = add_uv_ghost_cells(u, v)
-        u_full = linear_extrapolate_over_cf_dynamic_thickness(u_ghost, h_static)
-        v_full = linear_extrapolate_over_cf_dynamic_thickness(v_ghost, h_static)
-
-        #various face-centred derivatives
-        dudx_ew, dudy_ew = ew_gradient(u_full)
-        dvdx_ew, dvdy_ew = ew_gradient(v_full)
-        dudx_ns, dudy_ns = ns_gradient(u_full)
-        dvdx_ns, dvdy_ns = ns_gradient(v_full)
-
-        #jax.debug.print("dudxew = {x}",x=dudx_ew)
-
-        #interpolate things onto face-centres
-        h_ghost = add_s_ghost_cells(h)
-        h_ew, h_ns = interp_cc_to_fc(h_ghost)
-        
-
-        mucoef = add_s_ghost_cells(mucoef)
-        mucoef_ew, mucoef_ns = interp_cc_to_fc(mucoef)
-        #jax.debug.print("mucoef_ew = {x}",x=mucoef_ew)
-
-        #calculate face-centred viscosity:
-        mu_ew = B * mucoef_ew * (dudx_ew**2 + dvdy_ew**2 + dudx_ew*dvdy_ew +\
-                    0.25*(dudy_ew+dvdx_ew)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        mu_ns = B * mucoef_ns * (dudx_ns**2 + dvdy_ns**2 + dudx_ns*dvdy_ns +\
-                    0.25*(dudy_ns+dvdx_ns)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        
-        #to account for calving front boundary condition, set effective viscosities
-        #of faces of all cells with zero thickness to zero:
-        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h_t<1e-2, 0, mu_ew[:, 1:]))
-        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h_t<1e-2, 0, mu_ew[:,:-1]))
-        mu_ns = mu_ns.at[1:, :].set(jnp.where(h_t<1e-2, 0, mu_ns[1:, :]))
-        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h_t<1e-2, 0, mu_ns[:-1,:]))
-        
-
-        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
-
-        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
-
-        x_mom_residual = visc_x + volume_x
-        y_mom_residual = visc_y + volume_y
-
-
-        #print(f"xmom {jnp.max(jnp.abs(x_mom_residual))}")
-        #print(f"ymom {jnp.max(jnp.abs(y_mom_residual))}")
-
-
-        #################################################################
-
-
-
-
-
-
-
-        ################ ADVECTION RESIDUALS #############################
-
-
-        hh = linear_extrapolate_over_cf_dynamic_thickness(h_ghost, h_static)
-        #hh = h.copy()
-
-        #print("-----------------------------")
-        #print("-----------------------------")
-        #print("-----------------------------")
-        #print("-----------------------------")
-        #print("-----------------------------")
-        #print("-----------------------------")
-        #print("-----------------------------")
-        #print("hstatic")
-        #print(h_static)
-        #print("h")
-        #print(h)
-        #print("hghost")
-        #print(h_ghost)
-        #print("hh")
-        #print(hh)
-
-        u_fc_ew, _ = interp_cc_to_fc(u_full)
-        _, v_fc_ns = interp_cc_to_fc(v_full)
-
-        u_signs = jnp.where(u_fc_ew>0, 1, -1)
-        v_signs = jnp.where(v_fc_ns>0, 1, -1)
-
-
-        ##face-centred values according to first-order upwinding
-        h_fc_fou_ew = jnp.where(u_fc_ew>0, hh[1:-1,:-1], hh[1:-1, 1:])
-        h_fc_fou_ns = jnp.where(v_fc_ns>0, hh[1:, 1:-1], hh[-1:,1:-1])
-
-
-
-        flux_term = (u_fc_ew[:,1:]*h_fc_fou_ew[:,1:] - u_fc_ew[:,:-1]*h_fc_fou_ew[:,:-1])*dy*delta_t +\
-                    (v_fc_ns[:-1,:]*h_fc_fou_ns[:-1,:] - v_fc_ns[1:,:]*h_fc_fou_ns[1:,:])*dx*delta_t
-        #to keep calving front in same location, prevent any flux into or out of ice-free cells!
-        flux_term = jnp.where(h_t>1e-2, flux_term, 0)
-
-
-        adv_residual =  ((h - h_t) - source * delta_t)*dx*dy + flux_term
-
-        ##################################################################
-
-
-
-        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1), adv_residual.reshape(-1)
-
-    return compute_uvh_residuals
-
-
-def print_residual_things(residual, rhs, init_residual, i):
-    old_residual = residual
-    #residual = jnp.max(jnp.abs(-rhs))
-    residual = jnp.sqrt(jnp.sum(rhs**2))
-
-    if i==0:
-        init_residual = residual.copy()
-        print("Initial residual: {}".format(residual))
-    else:
-        print("residual: {}".format(residual))
-        print("Residual reduction factor: {}".format(old_residual/residual))
-    print("------")
-
-    return old_residual, residual, init_residual
-
-
-def make_newton_coupled_solver_function(ny, nx, dy, dx, C, n_iterations):
-
-    beta_eff = C.copy()
-
-    #functions for various things:
-    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
-    cc_gradient                                = cc_gradient_function(dy, dx)
-    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
-    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
-    get_uvh_residuals = compute_uvh_residuals_function(ny, nx, dy, dx,\
-                                                       beta_eff,\
-                                                       interp_cc_to_fc,\
-                                                       ew_gradient, ns_gradient,\
-                                                       cc_gradient,\
-                                                       add_uv_ghost_cells,\
-                                                       add_cont_ghost_cells)
-
-    #############
-    #setting up bvs and coords for a single block of the jacobian
-    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
-                                                                                  periodic_x=False)
-
-    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
-    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
-    mask = (i_coordinate_sets>=0)
-
-
-    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
-                                                        basis_vectors,\
-                                                        i_coordinate_sets,\
-                                                        j_coordinate_sets,\
-                                                        mask,\
-                                                        3,
-                                                        active_indices=(0,1,2)
-                                                       )
-    #sparse_jacrev = jax.jit(sparse_jacrev)
-
-
-    i_coordinate_sets = i_coordinate_sets[mask]
-    j_coordinate_sets = j_coordinate_sets[mask]
-    #############
-
-    coords = jnp.stack([
-        jnp.concatenate(
-           [i_coordinate_sets,           i_coordinate_sets,           i_coordinate_sets,\
-            i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),\
-            i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx)]
-                       ),\
-        jnp.concatenate(
-           [j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
-            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
-            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx)]
-                       )
-                       ])
-
-   
-    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*3, ny*nx*3),\
-                                                              ksp_type="gmres",\
-                                                              preconditioner="hypre",\
-                                                              precondition_only=False,\
-                                                              monitor_ksp=False)
-
-
-
-    comm = PETSc.COMM_WORLD
-    size = comm.Get_size()
-    
-    #@custom_vjp
-    def solver(q, u_trial, v_trial, h_now, accm, delta_t):
-        u_trial = jnp.where(h_now>1e-10, u_trial, 0)
-        v_trial = jnp.where(h_now>1e-10, v_trial, 0)
-
-        h_1d = h_now.copy().reshape(-1)
-        u_1d = u_trial.copy().reshape(-1)
-        v_1d = v_trial.copy().reshape(-1)
-
-        residual = jnp.inf
-        init_res = 0
-
-        old_residual = jnp.inf
-
-        #get_uvh_residuals(u_1d, v_1d, h_1d, q, h_now, accm, delta_t)
-
-        for i in range(n_iterations):
-
-            dRu_du, dRv_du, dRh_du, \
-            dRu_dv, dRv_dv, dRh_dv, \
-            dRu_dh, dRv_dh, dRh_dh = sparse_jacrev(get_uvh_residuals, \
-                                                  (u_1d, v_1d, h_1d, q, h_now, accm, delta_t)
-                                                          )
-
-            nz_jac_values = jnp.concatenate([dRu_du[mask], dRu_dv[mask], dRu_dh[mask],\
-                                             dRv_du[mask], dRv_dv[mask], dRv_dh[mask],\
-                                             dRh_du[mask], dRh_dv[mask], dRh_dh[mask]])
-
-          
-            #full_jac = jnp.zeros((ny*nx*3, ny*nx*3))
-            #full_jac = full_jac.at[coords[0,:], coords[1,:]].set(nz_jac_values)
-            #print(full_jac)
-            #raise
-
-
-            rhs = -jnp.concatenate(get_uvh_residuals(u_1d, v_1d, h_1d, q, h_now, accm, delta_t))
-
-            #print(jnp.max(rhs))
-            #raise
-
-            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
-            
-            
-
-
-            du = la_solver(nz_jac_values, rhs)
-
-
-
-            #residual = jnp.sqrt(jnp.sum(rhs**2))
-            #print(residual)
-            #print(old_residual/residual)
-            #old_residual = residual.copy()
-
-            #print(f"!!! {jnp.max(jnp.abs(du))}")
-            #
-
-            #iptr, j, values = scipy_coo_to_csr(nz_jac_values, coords,\
-            #                               (ny*nx*3, ny*nx*3), return_decomposition=True)
-
-            #A = PETSc.Mat().createAIJ(size=(ny*nx*3, ny*nx*3), \
-            #                          csr=(iptr.astype(np.int32), j.astype(np.int32), values),\
-            #                          comm=comm)
-            #
-            #b = PETSc.Vec().createWithArray(du, comm=comm)
-
-            #jdu = jnp.zeros_like(du)
-            #pjdu = PETSc.Vec().createWithArray(jdu, comm=comm)
-            #A.mult(b, pjdu)
-
-            #print(f"Jdeltau = {jnp.max(jnp.abs(jdu))}")
-
-            #
-            #print(f"R = {jnp.max(jnp.abs(jdu - rhs))}")
-
-
-
-
-
-
-
-            u_1d = u_1d+du[:(ny*nx)]
-            v_1d = v_1d+du[(ny*nx):(2*ny*nx)]
-            h_1d = h_1d+du[(2*ny*nx):]
-
-
-
-
-        res_final = jnp.max(jnp.abs(jnp.concatenate(
-                                    get_uvh_residuals(u_1d, v_1d, h_1d,  q, h_now, accm, delta_t)
-                                                   )
-                                   )
-                           )
-        print("----------")
-        print("Final residual: {}".format(res_final))
-        print("Total residual reduction factor: {}".format(init_res/res_final))
-        print("----------")
-        
-        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx)), h_1d.reshape((ny,nx))
-
-    return solver
-
-
-
-
-def make_quasi_newton_coupled_solver_function(ny, nx, dy, dx, C, n_iterations):
-
-    beta_eff = C.copy()
-
-    #functions for various things:
-    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
-    cc_gradient                                = cc_gradient_function(dy, dx)
-    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
-    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
-    get_uvh_residuals = compute_uvh_residuals_function_lr(ny, nx, dy, dx,\
-                                                          beta_eff,\
-                                                          interp_cc_to_fc,\
-                                                          ew_gradient, ns_gradient,\
-                                                          cc_gradient,\
-                                                          add_uv_ghost_cells,\
-                                                          add_cont_ghost_cells)
-
-    #############
-    #setting up bvs and coords for a single block of the jacobian
-    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
-                                                                                  periodic_x=False)
-
-    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
-    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
-    mask = (i_coordinate_sets>=0)
-
-
-    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
-                                                        basis_vectors,\
-                                                        i_coordinate_sets,\
-                                                        j_coordinate_sets,\
-                                                        mask,\
-                                                        3,
-                                                        active_indices=(0,1,2)
-                                                       )
-    #sparse_jacrev = jax.jit(sparse_jacrev)
-
-
-    i_coordinate_sets = i_coordinate_sets[mask]
-    j_coordinate_sets = j_coordinate_sets[mask]
-    #############
-
-    coords = jnp.stack([
-        jnp.concatenate(
-           [i_coordinate_sets,           i_coordinate_sets,           i_coordinate_sets,\
-            i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),   i_coordinate_sets+(ny*nx),\
-            i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx), i_coordinate_sets+(2*ny*nx)]
-                       ),\
-        jnp.concatenate(
-           [j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
-            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx),\
-            j_coordinate_sets, j_coordinate_sets+(ny*nx), j_coordinate_sets+(2*ny*nx)]
-                       )
-                       ])
-
-   
-    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*3, ny*nx*3),\
-                                                              ksp_type="bcgs",\
-                                                              preconditioner="hypre",\
-                                                              precondition_only=False,\
-                                                              monitor_ksp=False)
-
-
-
-    #@custom_vjp
-    def solver(q, u_trial, v_trial, h_now, accm, delta_t):
-        u_trial = jnp.where(h_now>1e-10, u_trial, 0)
-        v_trial = jnp.where(h_now>1e-10, v_trial, 0)
-
-        h_1d = h_now.copy().reshape(-1)
-        u_1d = u_trial.copy().reshape(-1)
-        v_1d = v_trial.copy().reshape(-1)
-
-        residual = jnp.inf
-        init_res = 0
-
-
-        viscosity = fc_viscosity(u_trial, v_trial, q, h_now)
-
-
-        for i in range(n_iterations):
-
-            dRu_du, dRv_du, dRh_du, \
-            dRu_dv, dRv_dv, dRh_dv, \
-            dRu_dh, dRv_dh, dRh_dh = sparse_jacrev(get_uvh_residuals_lr, \
-                                        (u_1d, v_1d, h_1d, viscosity, h_now, accm, delta_t)
-                                                          )
-
-            nz_jac_values = jnp.concatenate([dRu_du[mask], dRu_dv[mask], dRu_dh[mask],\
-                                             dRv_du[mask], dRv_dv[mask], dRv_dh[mask],\
-                                             dRh_du[mask], dRh_dv[mask], dRh_dh[mask]])
-
-          
-            #full_jac = jnp.zeros((ny*nx*3, ny*nx*3))
-            #full_jac = full_jac.at[coords[0,:], coords[1,:]].set(nz_jac_values)
-            #print(full_jac)
-            #raise
-
-
-            rhs = -jnp.concatenate(get_uvh_residuals_lr(u_1d, v_1d, h_1d,
-                                                        viscosity, h_now, accm, delta_t
-                                                        )
-                                   )
-
-            #print(jnp.max(rhs))
-            #raise
-
-            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
-
-
-            du = la_solver(nz_jac_values, rhs)
-
-            u_1d = u_1d+du[:(ny*nx)]
-            v_1d = v_1d+du[(ny*nx):(2*ny*nx)]
-            h_1d = h_1d+du[(2*ny*nx):]
-
-
-        res_final = jnp.max(jnp.abs(jnp.concatenate(
-                                get_uvh_residuals_lr(u_1d, v_1d, h_1d, 
-                                                     viscosity, h_now, accm, delta_t
-                                                    )
-                                                   )
-                                   )
-                           )
-
-        print("----------")
-        print("Final residual: {}".format(res_final))
-        print("Total residual reduction factor: {}".format(init_res/res_final))
-        print("----------")
-        
-        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx)), h_1d.reshape((ny,nx))
-
-    return solver
-
-
-
-
-def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
-                                                    C, n_iterations):
-
-    beta_eff = C.copy()
-
-    #functions for various things:
-    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
-    cc_gradient                                = cc_gradient_function(dy, dx)
-    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
-    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
-    #add_uv_ghost_cells                         = apply_scalar_ghost_cells_to_vector(
-    #                                                add_ghost_cells_periodic_dirichlet_function(ny,nx)
-    #                                             )
-    extrapolate_over_cf                        = linear_extrapolate_over_cf_dynamic_thickness
-
-    get_u_v_residuals = compute_u_v_residuals_function(ny, nx, dy, dx,\
-                                                       beta_eff,\
-                                                       interp_cc_to_fc,\
-                                                       ew_gradient, ns_gradient,\
-                                                       cc_gradient,\
-                                                       add_uv_ghost_cells,\
-                                                       add_cont_ghost_cells,\
-                                                       extrapolate_over_cf)
-    
-
-    #############
-    #setting up bvs and coords for a single block of the jacobian
-    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
-                                                                                  periodic_x=False)
-
-    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
-    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
-    mask = (i_coordinate_sets>=0)
-
-
-    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
-                                                        basis_vectors,\
-                                                        i_coordinate_sets,\
-                                                        j_coordinate_sets,\
-                                                        mask,\
-                                                        2,
-                                                        active_indices=(0,1)
-                                                       )
-    #sparse_jacrev = jax.jit(sparse_jacrev)
-
-
-    i_coordinate_sets = i_coordinate_sets[mask]
-    j_coordinate_sets = j_coordinate_sets[mask]
-    #############
-
-    coords = jnp.stack([
-                    jnp.concatenate(
-                                [i_coordinate_sets,         i_coordinate_sets,\
-                                 i_coordinate_sets+(ny*nx), i_coordinate_sets+(ny*nx)]
-                                   ),\
-                    jnp.concatenate(
-                                [j_coordinate_sets, j_coordinate_sets+(ny*nx),\
-                                 j_coordinate_sets, j_coordinate_sets+(ny*nx)]
-                                   )
-                       ])
-
-   
-    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*2, ny*nx*2),\
-                                                              ksp_type="bcgs",\
-                                                              preconditioner="hypre",\
-                                                              precondition_only=False)
-
-
-
-    @custom_vjp
-    def solver(q, u_trial, v_trial, h):
-        u_trial = jnp.where(h>1e-10, u_trial, 0)
-        v_trial = jnp.where(h>1e-10, v_trial, 0)
-
-        h_1d = h.copy().reshape(-1)
-        u_1d = u_trial.copy().reshape(-1)
-        v_1d = v_trial.copy().reshape(-1)
-
-        residual = jnp.inf
-        init_res = 0
-
-        for i in range(n_iterations):
-
-            dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
-                                                           (u_1d, v_1d, q, h_1d)
-                                                          )
-
-            nz_jac_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-                                             dJv_du[mask], dJv_dv[mask]])
-
-            #full_jac = jnp.zeros((ny*nx*2, ny*nx*2))
-            #full_jac = full_jac.at[coords[0,:], coords[1,:]].set(nz_jac_values)
-            #print(full_jac)
-            #raise
-
-
-            rhs = -jnp.concatenate(get_u_v_residuals(u_1d, v_1d, q, h_1d))
-
-            #print(jnp.max(rhs))
-            #raise
-
-            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
-
-
-            du = la_solver(nz_jac_values, rhs)
-
-            u_1d = u_1d+du[:(ny*nx)]
-            v_1d = v_1d+du[(ny*nx):]
-
-
-        res_final = jnp.max(jnp.abs(jnp.concatenate(
-                                    get_u_v_residuals(u_1d, v_1d, q, h_1d)
-                                                   )
-                                   )
-                           )
-        print("----------")
-        print("Final residual: {}".format(res_final))
-        print("Total residual reduction factor: {}".format(init_res/res_final))
-        print("----------")
-        
-        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx))
-
-
-
-    def solver_fwd(q, u_trial, v_trial, h):
-        u_trial = jnp.where(h>1e-10, u_trial, 0)
-        v_trial = jnp.where(h>1e-10, v_trial, 0)
-        
-        u, v = solver(q, u_trial, v_trial, h)
-
-        dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
-                                                       (u.reshape(-1), v.reshape(-1), q, h.reshape(-1))
-                                                      )
-        dJ_dvel_nz_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-                                             dJv_du[mask], dJv_dv[mask]])
-
-
-        fwd_residuals = (u, v, dJ_dvel_nz_values, q, h)
-        #fwd_residuals = (u, v, q)
-
-        return (u, v), fwd_residuals
-
-
-    def solver_bwd(res, cotangent):
-        
-        u, v, dJ_dvel_nz_values, q, h = res
-        #u, v, q = res
-
-        u_bar, v_bar = cotangent
-        
-
-        #dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
-        #                                               (u.reshape(-1), v.reshape(-1), q)
-        #                                              )
-        #dJ_dvel_nz_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-        #                                     dJv_du[mask], dJv_dv[mask]])
-
-
-        lambda_ = la_solver(dJ_dvel_nz_values,
-                            -jnp.concatenate([u_bar, v_bar]),
-                            transpose=True)
-
-        lambda_u = lambda_[:(ny*nx)]
-        lambda_v = lambda_[(ny*nx):]
-
-
-        #phi_bar = (dG/dphi)^T lambda
-        _, pullback_function = jax.vjp(get_u_v_residuals,
-                                         u.reshape(-1), v.reshape(-1), q, h.reshape(-1)
-                                      )
-        _, _, mu_bar = pullback_function((lambda_u, lambda_v))
-        
-#        #bwd has to return a tuple of cotangents for each primal input
-#        #of solver, so have to return this 1-tuple:
-#        return (mu_bar.reshape((ny, nx)), )
-
-        #I wonder if I can get away with just returning None for u_trial_bar, v_trial_bar, h_bar...
-        return (mu_bar.reshape((ny, nx)), None, None, None)
-
-
-    solver.defvjp(solver_fwd, solver_bwd)
-
-    return solver
-
-
-
-##NOTE: make everything linear by changing to 1
-nvisc = c.GLEN_N
-#nvisc = 1.001
-
-A = c.A_COLD
-B = 0.5 * (A**(-1/nvisc))
 
 
 def ice_shelf():
@@ -1115,16 +1109,16 @@ v_init = jnp.zeros_like(b)
 n_iterations = 10
 
 
-timestep = 25 #year
+timestep = 10 #year
 
 
 
 def expl_ts():
-    vel_solver = make_newton_velocity_solver_function_custom_vjp(nr, nc,
+    vel_solver = make_newton_velocity_solver_function_custom_vjp_dynamic_thk(nr, nc,
                                                              delta_y,
                                                              delta_x,
-                                                             C,
-                                                             n_iterations)
+                                                             C, b,
+                                                             n_iterations, mucoef_0)
     
     thickness_update = make_advect_scalar_field_fou_function(nc, nr, 
                                                              delta_x, delta_y,
@@ -1149,14 +1143,14 @@ def impl_ts():
     uvh_solver = make_newton_coupled_solver_function(nr, nc,
                                                      delta_y,
                                                      delta_x,
-                                                     C,
-                                                     n_iterations)
+                                                     C, b,
+                                                     n_iterations, mucoef_0)
     
-    vel_solver = make_newton_velocity_solver_function_custom_vjp(nr, nc,
+    vel_solver = make_newton_velocity_solver_function_custom_vjp_dynamic_thk(nr, nc,
                                                              delta_y,
                                                              delta_x,
-                                                             C,
-                                                             n_iterations)
+                                                             C, b,
+                                                             n_iterations, mucoef_0)
    
     ui = u_init.copy()
     vi = v_init.copy()
