+++ scratch_test/expt/SOA/ice_shelf/bank/soa_test.py
@@ -17,7 +17,6 @@ from plotting_stuff import show_vel_field, make_gif, show_damage_field,\
                            create_imageio_gif, create_webp_from_pngs, create_gif_global_palette
 
 
-
 #3rd party
 from petsc4py import PETSc
 #from mpi4py import MPI
@@ -71,7 +70,7 @@ def create_sparse_petsc_la_solver_with_custom_vjp(coordinates, jac_shape,\
         
         #set ksp iterations
         opts = PETSc.Options()
-        opts['ksp_max_it'] = 20
+        opts['ksp_max_it'] = 40
         if monitor_ksp:
             opts['ksp_monitor'] = None
         opts['ksp_rtol'] = 1e-20
@@ -392,12 +391,12 @@ def apply_scalar_ghost_cells_to_vector(scalar_ghost_function):
 
 def binary_erosion(boolean_array):
     # 3x3 cross-shaped structuring element (4-connectivity)
-    #kernel = jnp.array([[1,1,1],
-    #                    [1,1,1],
-    #                    [1,1,1]], dtype=jnp.bool_)
-    kernel = jnp.array([[0,1,0],
+    kernel = jnp.array([[1,1,1],
                         [1,1,1],
-                        [0,1,0]], dtype=jnp.bool_)
+                        [1,1,1]], dtype=jnp.bool_)
+    #kernel = jnp.array([[0,1,0],
+    #                    [1,1,1],
+    #                    [0,1,0]], dtype=jnp.bool_)
 
     kernel = kernel.astype(jnp.float64)
 
@@ -411,7 +410,7 @@ def binary_erosion(boolean_array):
             padding='SAME',
             dimension_numbers=('NCHW','OIHW','NCHW')
         )
-        return (out[0,0] > 3).astype(jnp.bool_)
+        return (out[0,0] > 8).astype(jnp.bool_)
 
     return erode_once(boolean_array.astype(jnp.float64))
 
@@ -443,24 +442,25 @@ def binary_dilation(boolean_array):
 #        return cc_field
 #    return jax.jit(extrapolate_over_cf)
 
+def safe_neighbours(cc_field):
+    # cc_field shape (ny, nx)
+    up    = jnp.pad(cc_field, ((0,1),(0,0)))[1: , :]   # shifted up
+    down  = jnp.pad(cc_field, ((1,0),(0,0)))[:-1, :]   # shifted down
+    left  = jnp.pad(cc_field, ((0,0),(0,1)))[:, 1: ]   # shifted left
+    right = jnp.pad(cc_field, ((0,0),(1,0)))[:, :-1]   # shifted right
+    return up, down, left, right
+
 def extrapolate_over_cf_function(thk):
     
     cf_adjacent_zero_ice_cells = (thk==0) & binary_dilation(thk>0)
 
     ice_mask = (thk>0)
 
-    ice_mask_shift_up    = jnp.roll(ice_mask, -1, axis=0)
-    ice_mask_shift_down  = jnp.roll(ice_mask,  1, axis=0)
-    ice_mask_shift_left  = jnp.roll(ice_mask, -1, axis=1)
-    ice_mask_shift_right = jnp.roll(ice_mask,  1, axis=1)
-    
+    ice_mask_shift_up, ice_mask_shift_down, ice_mask_shift_left, ice_mask_shift_right = safe_neighbours(ice_mask)
 
     def extrapolate_over_cf(cc_field):
 
-        u_shift_up    = jnp.roll(cc_field, -1, axis=0)
-        u_shift_down  = jnp.roll(cc_field,  1, axis=0)
-        u_shift_left  = jnp.roll(cc_field, -1, axis=1)
-        u_shift_right = jnp.roll(cc_field,  1, axis=1)
+        u_shift_up, u_shift_down, u_shift_left, u_shift_right = safe_neighbours(cc_field)
         
         neighbour_values = jnp.stack([
             jnp.where(ice_mask_shift_up   ==1, u_shift_up,    0),
@@ -469,19 +469,43 @@ def extrapolate_over_cf_function(thk):
             jnp.where(ice_mask_shift_right==1, u_shift_right, 0),
         ])
         
-        neighbour_counts = jnp.stack([
-            (ice_mask_shift_up   ==1).astype(int),
-            (ice_mask_shift_down ==1).astype(int),
-            (ice_mask_shift_left ==1).astype(int),
-            (ice_mask_shift_right==1).astype(int),
-        ])
+        #neighbour_counts = jnp.stack([
+        #    (ice_mask_shift_up   ==1).astype(int),
+        #    (ice_mask_shift_down ==1).astype(int),
+        #    (ice_mask_shift_left ==1).astype(int),
+        #    (ice_mask_shift_right==1).astype(int),
+        #]).sum(axis=0)
+        #neighbour_counts = jnp.where(neighbour_counts == 0, 1, neighbour_counts)
+        #
+
+        ##NOTE: STEPH!!
+        ##Including this factor of 2 screws the gradient as computed by the HVP, but fixes
+        ##the gradient as computed by the adjoint models. Make of that what you will...
+        ##u_extrap_boundary = 2 * neighbour_values.sum(axis=0) / neighbour_counts.sum(axis=0)
+        #u_extrap_boundary = neighbour_values.sum(axis=0) / neighbour_counts
+        #u_extrap_boundary = jnp.where(neighbour_counts==0, 0, u_extrap_boundary)
+        ##Think about it... 
         
-        u_extrap_boundary = 2 * neighbour_values.sum(axis=0) / neighbour_counts.sum(axis=0)
-        #Think about it... 
+        u_extrap_boundary = jnp.take_along_axis(neighbour_values,
+                                                jnp.argmax(jnp.abs(neighbour_values), axis=0)[None, ...],
+                                                axis=0)[0]
 
         return cc_field + u_extrap_boundary*cf_adjacent_zero_ice_cells.astype(jnp.float64)
 
     return extrapolate_over_cf
+
+
+#test_thk = jnp.ones((10,10))
+#test_thk = test_thk.at[:,-1].set(0)
+#test_array = jnp.ones_like(test_thk)*jnp.linspace(0,8,10)
+#test_array = test_array.at[:,-1].set(0)
+#print(test_array)
+#
+#e = extrapolate_over_cf_function(test_thk)
+#
+#ta_ex = e(test_array)
+#print(ta_ex)
+#raise
 
 @jax.jit
 def double_dot_contraction(A, B):
@@ -835,6 +859,8 @@ def compute_u_v_residuals_function(ny, nx, dy, dx, \
 
         #obvs not going to do anything in the no-cf case
         u = extrp_over_cf(u)
+        #jax.debug.print("{x}",x=u[:,-3:])
+
         v = extrp_over_cf(v)
         #momentum_term
         u, v = add_uv_ghost_cells(u, v)
@@ -1039,6 +1065,9 @@ def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
 
     @custom_vjp
     def solver(q, u_trial, v_trial):
+        u_trial = jnp.where(h>1e-10, u_trial, 0)
+        v_trial = jnp.where(h>1e-10, v_trial, 0)
+
         u_1d = u_trial.copy().reshape(-1)
         v_1d = v_trial.copy().reshape(-1)
 
@@ -1084,6 +1113,9 @@ def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
 
 
     def solver_fwd(q, u_trial, v_trial):
+        u_trial = jnp.where(h>1e-10, u_trial, 0)
+        v_trial = jnp.where(h>1e-10, v_trial, 0)
+        
         u, v = solver(q, u_trial, v_trial)
 
         dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
@@ -1094,7 +1126,7 @@ def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
 
 
         fwd_residuals = (u, v, dJ_dvel_nz_values, q)
-#        fwd_residuals = (u, v, q)
+        #fwd_residuals = (u, v, q)
 
         return (u, v), fwd_residuals
 
@@ -1117,10 +1149,6 @@ def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
         lambda_ = la_solver(dJ_dvel_nz_values,
                             -jnp.concatenate([u_bar, v_bar]),
                             transpose=True)
-
-        #NOTE: Ok, so if you stop the gradients through lambda, it gets rid of the
-        #massive gradients. But it also gets rid of the interesting information...
-        #lambda_ = jax.lax.stop_gradient(lambda_)
 
         lambda_u = lambda_[:(ny*nx)]
         lambda_v = lambda_[(ny*nx):]
@@ -1307,6 +1335,9 @@ def forward_adjoint_and_second_order_adjoint_solvers(ny, nx, dy, dx, h, C, n_ite
     #picard_solver = make_picard_solver(ny, nx, sparse_jacrev, mask, la_solver, get_u_v_residuals, fc_visc)
 
     def solve_fwd_problem(q, u_trial, v_trial):
+        u_trial = jnp.where(h>1e-10, u_trial, 0)
+        v_trial = jnp.where(h>1e-10, v_trial, 0)
+        
         #u, v = newton_solver(u_trial, v_trial, get_u_v_residuals, n_iterations, (q,), coords)
         u, v = newton_solver(u_trial, v_trial, get_u_v_residuals, n_iterations, (q,))
         return u.reshape((ny,nx)), v.reshape((ny,nx))
@@ -1358,6 +1389,16 @@ def forward_adjoint_and_second_order_adjoint_solvers(ny, nx, dy, dx, h, C, n_ite
                                linear_ssa_residuals, 1, (mu_bar_ew, mu_bar_ns, rhs))
 
         mu_bar = cc_viscosity(q, u, v)
+
+        #plt.imshow(mu_bar)
+        #plt.colorbar()
+        #plt.show()
+
+        #plt.imshow(mu_bar_ns)
+        #plt.colorbar()
+        #plt.show()
+
+
         #calculate gradient
         dJdq = jax.grad(functional, argnums=2)(*argz) \
                - mu_bar * h * double_dot_contraction(cc_vector_field_gradient(lx.reshape(-1), ly.reshape(-1)),
@@ -1456,7 +1497,7 @@ def ice_shelf():
     lx = 150_000
     ly = 200_000
     
-    resolution = 2000 #m
+    resolution = 4000 #m
     
     nr = int(ly/resolution)
     nc = int(lx/resolution)
@@ -1472,7 +1513,7 @@ def ice_shelf():
 
     thk_profile = 500# - 300*x/lx
     thk = jnp.zeros((nr, nc))+thk_profile
-    thk = thk.at[:, -1:].set(0)
+    thk = thk.at[:, -2:].set(0)
     
     b = jnp.zeros_like(thk)-600
     
@@ -1504,6 +1545,10 @@ n_iterations = 10
 mask = jnp.where(thk>0,1,0)
 mask = binary_erosion(mask)
 mask = binary_erosion(mask)
+mask = binary_erosion(mask)
+mask = mask.astype(int)
+#print(mask)
+#raise
 
 def functional(v_field_x, v_field_y, q):
     #NOTE: for things like this, you need to ensure the value isn't
@@ -1581,6 +1626,7 @@ def calculate_hvp_via_soa():
 
     print("solving second-order adjoint problem:")
     pert_dir = gradient.copy()/(jnp.linalg.norm(gradient)*10)
+    pert_dir = pert_dir.at[:,-4:].set(0)
     hvp = soa_solver(q, u_out, v_out, lx, ly, pert_dir, functional)
     
     #plt.imshow(hvp[:,:], vmin=-3, vmax=3, cmap="twilight_shifted")
@@ -1610,19 +1656,19 @@ def calculate_hvp_via_ad():
 
     gradient = get_grad(q)
     
-    #plt.imshow(gradient[:,:])
-    #plt.title("gradient via ad")
-    #plt.colorbar()
-    #plt.imshow(jnp.where(C>1e10, 1, jnp.nan), cmap="Grays", alpha=0.5)
-    #plt.show()
+    plt.imshow(gradient[:,:])
+    plt.title("gradient via ad")
+    plt.colorbar()
+    plt.imshow(jnp.where(C>1e10, 1, jnp.nan), cmap="Grays", alpha=0.5)
+    plt.show()
 
     #plt.plot(gradient[25,:])
     #plt.ylim((-600,600))
     #plt.show()
 
     
-    pert_dir = gradient / (jnp.linalg.norm(gradient)*10)
-
+    pert_dir = gradient.copy() / (jnp.linalg.norm(gradient)*10)
+    pert_dir = pert_dir.at[:,-4:].set(0)
 
     ##finite diff hvp for comparison
     ##plt.imshow(p)
@@ -1703,7 +1749,7 @@ def calculate_hvp_via_ad():
 
 
 
-#calculate_hvp_via_ad()
+calculate_hvp_via_ad()
 calculate_hvp_via_soa()
