+++ scratch_test/expt/SOA/ice_shelf/soa_test.py
@@ -15,7 +15,12 @@ import constants_years as c
 from plotting_stuff import show_vel_field, make_gif, show_damage_field,\
                            create_gif_from_png_fps, create_high_quality_gif_from_pngfps,\
                            create_imageio_gif, create_webp_from_pngs, create_gif_global_palette
+from grid import *
 
+sys.path.insert(1, "../../../../solvers/")
+from linear_solvers import create_sparse_petsc_la_solver_with_custom_vjp
+from nonlinear_solvers import forward_adjoint_and_second_order_adjoint_solvers,\
+        make_newton_velocity_solver_function_custom_vjp
 
 #3rd party
 from petsc4py import PETSc
@@ -42,1462 +47,12 @@ import matplotlib.cm as cm
 np.set_printoptions(precision=1, suppress=True, linewidth=np.inf, threshold=np.inf)
 jax.config.update("jax_enable_x64", True)
 
-def create_sparse_petsc_la_solver_with_custom_vjp(coordinates, jac_shape,\
-                                    ksp_type='gmres', preconditioner='hypre',\
-                                    precondition_only=False, monitor_ksp=False):
-
-    comm = PETSc.COMM_WORLD
-    size = comm.Get_size()
-    
-    def construct_ab(values, b, transpose):
-        if transpose:
-            iptr, j, values = scipy_coo_to_csr(values, coordinates[::-1,:],\
-                                           jac_shape, return_decomposition=True)
-        else:
-            iptr, j, values = scipy_coo_to_csr(values, coordinates,\
-                                           jac_shape, return_decomposition=True)
-        #rows_local = int(jac_shape[0] / size)
-
-        A = PETSc.Mat().createAIJ(size=jac_shape, \
-                                  csr=(iptr.astype(np.int32), j.astype(np.int32), values),\
-                                  comm=comm)
-        
-        b = PETSc.Vec().createWithArray(b, comm=comm)
-        return A, b
-
-    
-    def create_solver_object(A):
-        
-        #set ksp iterations
-        opts = PETSc.Options()
-        opts['ksp_max_it'] = 40
-        if monitor_ksp:
-            opts['ksp_monitor'] = None
-        opts['ksp_rtol'] = 1e-20
-        
-        # Create a linear solver
-        ksp = PETSc.KSP().create()
-        ksp.setType(ksp_type)
-
-        ksp.setOperators(A)
-        ksp.setFromOptions()
-
-        if preconditioner is not None:
-            #assessing if preconditioner is doing anything:
-            #print((A*x - b).norm())
-
-            if preconditioner == 'hypre':
-                pc = ksp.getPC()
-                pc.setType('hypre')
-                pc.setHYPREType('boomeramg')
-            else:
-                pc = ksp.getPC()
-                pc.setType(preconditioner)
-        
-            return ksp, pc
-        else:
-            return ksp, None
-
-
-    @partial(jax.custom_vjp, nondiff_argnums=(2,))
-    def petsc_sparse_la_solver(values, b, transpose=False):
-    
-        A, b = construct_ab(values, b, transpose)
-        x = b.duplicate()
-        
-        ksp, pc = create_solver_object(A)
-
-        if precondition_only:
-            pc.apply(b, x)
-        else:
-            ksp.solve(b, x)
-        
-        x_jnp = jnp.array(x.getArray())
-
-        return x_jnp
-
-    
-    def la_solver_fwd(values, b, transpose=False):
-        solution = petsc_sparse_la_solver(values, b, transpose=transpose)
-        return solution, (values, b, solution)
-
-
-    #NOTE: the nondiff_argnums=(2,) thing shunts the transpose ragument to the front.
-    def linear_solve_bwd(transpose, res, x_bar):
-        #NOTE: The sign convention here is correct, despite what people say...
-        #It just follows the documentation rather than "textbook" versions.
-        values, b, x = res
-
-        lambda_ = petsc_sparse_la_solver(values, -x_bar, transpose=True)
-
-        b_bar = -lambda_.reshape(b.shape) #ensure same shape as input b.
-
-        #sparse version of jnp.outer(x,lambda_)
-        #TODO: CHECK WHICH WAY ROUND THESE COORDS GO. SHOULD BE RIGHT IF THEY ARE IJ!
-        values_bar = x[coordinates[1]] * lambda_[coordinates[0]]
-
-        return values_bar, b_bar
-
-    petsc_sparse_la_solver.defvjp(la_solver_fwd, linear_solve_bwd)
-
-    return petsc_sparse_la_solver
-
-def create_sparse_petsc_la_solver_test(jac_shape, ksp_type='gmres', preconditioner='hypre',\
-                                    precondition_only=False, monitor_ksp=False):
-
-    comm = PETSc.COMM_WORLD
-    size = comm.Get_size()
-    
-    def construct_ab(values, b, coordinates, transpose):
-        if transpose:
-            iptr, j, values = scipy_coo_to_csr(values, coordinates[::-1,:],\
-                                           jac_shape, return_decomposition=True)
-        else:
-            iptr, j, values = scipy_coo_to_csr(values, coordinates,\
-                                           jac_shape, return_decomposition=True)
-        #rows_local = int(jac_shape[0] / size)
-
-        A = PETSc.Mat().createAIJ(size=jac_shape, \
-                                  csr=(iptr.astype(np.int32), j.astype(np.int32), values),\
-                                  comm=comm)
-        
-        b = PETSc.Vec().createWithArray(b, comm=comm)
-        return A, b
-
-    
-    def create_solver_object(A):
-        
-        #set ksp iterations
-        opts = PETSc.Options()
-        opts['ksp_max_it'] = 100
-        if monitor_ksp:
-            opts['ksp_monitor'] = None
-        opts['ksp_rtol'] = 1e-20
-        
-        # Create a linear solver
-        ksp = PETSc.KSP().create()
-        ksp.setType(ksp_type)
-
-        ksp.setOperators(A)
-        ksp.setFromOptions()
-
-        if preconditioner is not None:
-            #assessing if preconditioner is doing anything:
-            #print((A*x - b).norm())
-
-            if preconditioner == 'hypre':
-                pc = ksp.getPC()
-                pc.setType('hypre')
-                pc.setHYPREType('boomeramg')
-            else:
-                pc = ksp.getPC()
-                pc.setType(preconditioner)
-        
-            return ksp, pc
-        else:
-            return ksp, None
-
-
-    @partial(jax.custom_vjp, nondiff_argnums=(2,3))
-    def petsc_sparse_la_solver(values, b, coordinates, transpose=False):
-    
-        A, b = construct_ab(values, b, coordinates, transpose)
-        x = b.duplicate()
-        
-        ksp, pc = create_solver_object(A)
-
-        if precondition_only:
-            pc.apply(b, x)
-        else:
-            ksp.solve(b, x)
-        
-        x_jnp = jnp.array(x.getArray())
-
-        return x_jnp
-
-    
-    def la_solver_fwd(values, b, coordinates, transpose=False):
-        solution = petsc_sparse_la_solver(values, b, coordinates, transpose=transpose)
-        return solution, (values, b, solution)
-
-
-    #NOTE: the nondiff_argnums=(2,) thing shunts the transpose ragument to the front.
-    def linear_solve_bwd(transpose, coordinates, res, x_bar):
-        #NOTE: The sign convention here is correct, despite what people say...
-        #It just follows the documentation rather than "textbook" versions.
-        values, b, x = res
-
-        lambda_ = petsc_sparse_la_solver(values, -x_bar, coordinates, transpose=True)
-
-        b_bar = -lambda_.reshape(b.shape) #ensure same shape as input b.
-
-        #sparse version of jnp.outer(x,lambda_)
-        #TODO: CHECK WHICH WAY ROUND THESE COORDS GO. SHOULD BE RIGHT IF THEY ARE IJ!
-        values_bar = x[coordinates[1]] * lambda_[coordinates[0]]
-
-        return values_bar, b_bar
-
-    petsc_sparse_la_solver.defvjp(la_solver_fwd, linear_solve_bwd)
-
-    return petsc_sparse_la_solver
-
-def interp_cc_with_ghosts_to_fc_function(ny, nx):
-    def interp_cc_to_fc(var):
-        
-        var_ew = 0.5*(var[1:-1, 1:]+var[1:-1, :-1])
-        var_ns = 0.5*(var[:-1, 1:-1]+var[1:, 1:-1])
-
-        return var_ew, var_ns
-    return jax.jit(interp_cc_to_fc)
-
-
-def interp_cc_to_fc_function(ny, nx):
-
-    def interp_cc_to_fc(var):
-        
-        var_ew = jnp.zeros((ny, nx+1))
-        var_ew = var_ew.at[:, 1:-1].set(0.5*(var[:, 1:]+var[:, :-1]))
-        var_ew = var_ew.at[:, 0].set(var[:, 0])
-        var_ew = var_ew.at[:, -1].set(var[:, -1])
-
-        var_ns = jnp.zeros((ny+1, nx))
-        var_ns = var_ns.at[1:-1, :].set(0.5*(var[:-1, :]+var[1:, :]))
-        var_ns = var_ns.at[0, :].set(var[0, :])
-        var_ns = var_ns.at[-1, :].set(var[-1, :])
-
-        return var_ew, var_ns
-
-    return jax.jit(interp_cc_to_fc)
-
-
-def cc_gradient_function(dy, dx):
-
-    def cc_gradient(var):
-
-        dvar_dx = (0.5/dx) * (var[1:-1, 2:] - var[1:-1,:-2])
-        dvar_dy = (0.5/dy) * (var[:-2,1:-1] - var[2:, 1:-1])
-
-        return dvar_dx, dvar_dy
-
-    return jax.jit(cc_gradient)
-
-def fc_gradient_functions(dy, dx):
-
-    def ew_face_gradient(var):
-        
-        dvar_dx_ew = (var[1:-1, 1:] - var[1:-1, :-1])/dx
-
-        dvar_dy_ew = (var[:-2, 1:] + var[:-2, :-1] - var[2:, 1:] - var[2:, :-1])/(4*dy)
-        
-        return dvar_dx_ew, dvar_dy_ew
-    
-    def ns_face_gradient(var):
-        
-        dvar_dy_ns = (var[:-1, 1:-1]-var[1:, 1:-1])/dy
-
-        dvar_dx_ns = (var[:-1, 2:] + var[1:, 2:] - var[:-1, :-2] - var[1:, :-2])/(4*dx)
-        
-        return dvar_dx_ns, dvar_dy_ns
-    
-    return jax.jit(ew_face_gradient), jax.jit(ns_face_gradient)
-
-
-def add_ghost_cells_fcts(ny, nx):
-
-    def add_reflection_ghost_cells(u_int, v_int):
-
-        u = jnp.zeros((ny+2, nx+2))
-        u = u.at[1:-1,1:-1].set(u_int)
-        #edges
-        u = u.at[0, 1:-1].set( u[1, 1:-1])
-        u = u.at[-1,1:-1].set( u[-2,1:-1])
-        u = u.at[1:-1, 0].set(-u[1:-1, 1])
-        u = u.at[1:-1,-1].set(-u[1:-1,-2])
-        #corner points
-        u = u.at[0,0].set(-u[1,1])
-        u = u.at[-1,-1].set(-u[-2,-2])
-        u = u.at[-1,0].set(-u[-2,1])
-        u = u.at[0,-1].set(-u[1,-2])
-
-        
-        v = jnp.zeros((ny+2, nx+2))
-        v = v.at[1:-1,1:-1].set(v_int)
-        #edges
-        v = v.at[0, 1:-1].set(-v[1, 1:-1])
-        v = v.at[-1,1:-1].set(-v[-2,1:-1])
-        v = v.at[1:-1, 0].set( v[1:-1, 1])
-        v = v.at[1:-1,-1].set( v[1:-1,-2])
-        #corner points
-        v = v.at[0,0].set(-v[1,1])
-        v = v.at[-1,-1].set(-v[-2,-2])
-        v = v.at[-1,0].set(-v[-2,1])
-        v = v.at[0,-1].set(-v[1,-2])
-
-        return u, v
-
-    def add_continuation_ghost_cells(h_int):
-
-        h = jnp.zeros((ny+2, nx+2))
-        h = h.at[1:-1,1:-1].set(h_int)
-        #edges
-        h = h.at[0, 1:-1].set(h[1, 1:-1])
-        h = h.at[-1,1:-1].set(h[-2,1:-1])
-        h = h.at[1:-1, 0].set(h[1:-1, 1])
-        h = h.at[1:-1,-1].set(h[1:-1,-2])
-        #corner points
-        h = h.at[0,0].set(h[1,1])
-        h = h.at[-1,0].set(h[-2,1])
-        h = h.at[0,-1].set(h[1,-2])
-        h = h.at[-1,-1].set(h[-2,-2])
-
-        return h
-    
-    return jax.jit(add_reflection_ghost_cells), jax.jit(add_continuation_ghost_cells)
-    
-def add_ghost_cells_periodic_dirichlet_function(ny, nx):
-    def add_ghost_cells_periodic_x_dirchlet_y(u_int):
-
-        u = jnp.zeros((ny+2, nx+2))
-        u = u.at[1:-1,1:-1].set(u_int)
-        #left/right edges: periodic bcs
-        u = u.at[:, 0].set(u[:,-2])
-        u = u.at[:,-1].set(u[:, 1])
-        #top/bottom edges: dirchlet bcs
-        u = u.at[0, :].set(-u[1, :])
-        u = u.at[-1,:].set(-u[-2,:])
-
-        return u
-    return jax.jit(add_ghost_cells_periodic_x_dirchlet_y)
-
-def add_ghost_cells_periodic_continuation_function(ny, nx):
-    def add_ghost_cells_periodic_x_continuation_y(u_int):
-        u = jnp.zeros((ny+2, nx+2))
-        u = u.at[1:-1,1:-1].set(u_int)
-        #left/right edges: periodic bcs
-        u = u.at[:, 0].set(u[:,-2])
-        u = u.at[:,-1].set(u[:, 1])
-        #top/bottom edges: continuation bcs
-        u = u.at[0, :].set(u[1, :])
-        u = u.at[-1,:].set(u[-2,:])
-        return u
-    return jax.jit(add_ghost_cells_periodic_x_continuation_y)
-
-def apply_scalar_ghost_cells_to_vector(scalar_ghost_function):
-    def apply(u,v):
-        u = scalar_ghost_function(u)
-        v = scalar_ghost_function(v)
-        return u, v
-    return jnp.jit(apply)
-
-def binary_erosion(boolean_array):
-    # 3x3 cross-shaped structuring element (4-connectivity)
-    kernel = jnp.array([[1,1,1],
-                        [1,1,1],
-                        [1,1,1]], dtype=jnp.bool_)
-    #kernel = jnp.array([[0,1,0],
-    #                    [1,1,1],
-    #                    [0,1,0]], dtype=jnp.bool_)
-
-    kernel = kernel.astype(jnp.float64)
-
-    def erode_once(mask_float):
-        out = jax.lax.conv_general_dilated(
-            #shape (batch_size,channels,H,W)
-            mask_float[None, None, :, :],
-            #shape (out_chan,in_chan,H,W)
-            kernel[None, None, :, :],
-            window_strides=(1,1),
-            padding='SAME',
-            dimension_numbers=('NCHW','OIHW','NCHW')
-        )
-        return (out[0,0] > 8).astype(jnp.bool_)
-
-    return erode_once(boolean_array.astype(jnp.float64))
-
-def binary_dilation(boolean_array):
-    # 3x3 cross-shaped structuring element (4-connectivity)
-    kernel = jnp.array([[0,1,0],
-                        [1,1,1],
-                        [0,1,0]], dtype=jnp.bool_)
-
-    kernel = kernel.astype(jnp.float64)
-
-    def dilate_once(mask_float):
-        out = jax.lax.conv_general_dilated(
-            #shape (batch_size,channels,H,W)
-            mask_float[None, None, :, :],
-            #shape (out_chan,in_chan,H,W)
-            kernel[None, None, :, :],
-            window_strides=(1,1),
-            padding='SAME',
-            dimension_numbers=('NCHW','OIHW','NCHW')
-        )
-        return (out[0,0] > 0).astype(jnp.bool_)
-
-    return dilate_once(boolean_array.astype(jnp.float64))
-
-
-#def extrapolate_over_cf_function(thk):
-#    def extrapolate_over_cf(cc_field):
-#        return cc_field
-#    return jax.jit(extrapolate_over_cf)
-
-def safe_neighbours(cc_field):
-    # cc_field shape (ny, nx)
-    up    = jnp.pad(cc_field, ((0,1),(0,0)))[1: , :]   # shifted up
-    down  = jnp.pad(cc_field, ((1,0),(0,0)))[:-1, :]   # shifted down
-    left  = jnp.pad(cc_field, ((0,0),(0,1)))[:, 1: ]   # shifted left
-    right = jnp.pad(cc_field, ((0,0),(1,0)))[:, :-1]   # shifted right
-    return up, down, left, right
-
-def extrapolate_over_cf_function(thk):
-    
-    cf_adjacent_zero_ice_cells = (thk==0) & binary_dilation(thk>0)
-
-    ice_mask = (thk>0)
-
-    ice_mask_shift_up, ice_mask_shift_down, ice_mask_shift_left, ice_mask_shift_right = safe_neighbours(ice_mask)
-
-    def extrapolate_over_cf(cc_field):
-
-        u_shift_up, u_shift_down, u_shift_left, u_shift_right = safe_neighbours(cc_field)
-        
-        neighbour_values = jnp.stack([
-            jnp.where(ice_mask_shift_up   ==1, u_shift_up,    0),
-            jnp.where(ice_mask_shift_down ==1, u_shift_down,  0),
-            jnp.where(ice_mask_shift_left ==1, u_shift_left,  0),
-            jnp.where(ice_mask_shift_right==1, u_shift_right, 0),
-        ])
-        
-        #neighbour_counts = jnp.stack([
-        #    (ice_mask_shift_up   ==1).astype(int),
-        #    (ice_mask_shift_down ==1).astype(int),
-        #    (ice_mask_shift_left ==1).astype(int),
-        #    (ice_mask_shift_right==1).astype(int),
-        #]).sum(axis=0)
-        #neighbour_counts = jnp.where(neighbour_counts == 0, 1, neighbour_counts)
-        #
-
-        ##NOTE: STEPH!!
-        ##Including this factor of 2 screws the gradient as computed by the HVP, but fixes
-        ##the gradient as computed by the adjoint models. Make of that what you will...
-        ##u_extrap_boundary = 2 * neighbour_values.sum(axis=0) / neighbour_counts.sum(axis=0)
-        #u_extrap_boundary = neighbour_values.sum(axis=0) / neighbour_counts
-        #u_extrap_boundary = jnp.where(neighbour_counts==0, 0, u_extrap_boundary)
-        ##Think about it... 
-        
-        u_extrap_boundary = jnp.take_along_axis(neighbour_values,
-                                                jnp.argmax(jnp.abs(neighbour_values), axis=0)[None, ...],
-                                                axis=0)[0]
-
-        return cc_field + u_extrap_boundary*cf_adjacent_zero_ice_cells.astype(jnp.float64)
-
-    return extrapolate_over_cf
-
-
-#test_thk = jnp.ones((10,10))
-#test_thk = test_thk.at[:,-1].set(0)
-#test_array = jnp.ones_like(test_thk)*jnp.linspace(0,8,10)
-#test_array = test_array.at[:,-1].set(0)
-#print(test_array)
-#
-#e = extrapolate_over_cf_function(test_thk)
-#
-#ta_ex = e(test_array)
-#print(ta_ex)
-#raise
-
-@jax.jit
-def double_dot_contraction(A, B):
-    return A[:,:,0,0]*B[:,:,0,0] + A[:,:,1,0]*B[:,:,0,1] +\
-           A[:,:,0,1]*B[:,:,1,0] + A[:,:,1,1]*B[:,:,1,1]
-
-
-def cc_vector_field_gradient_function(ny, nx, dy, dx, cc_grad,
-                                      extrp_over_cf,
-                                      add_uv_ghost_cells):
-    def cc_vector_field_gradient(u, v):
-        u = u.reshape((ny, nx))
-        v = v.reshape((ny, nx))
-
-        u = extrp_over_cf(u)
-        v = extrp_over_cf(v)
-
-        u, v = add_uv_ghost_cells(u, v)
-
-        dudx, dudy = cc_grad(u)
-        dvdx, dvdy = cc_grad(v)
-
-        #dudx = jnp.where(thk>0, dudx, 0)
-        #dvdx = jnp.where(thk>0, dvdx, 0)
-        #dudy = jnp.where(thk>0, dudy, 0)
-        #dvdy = jnp.where(thk>0, dvdy, 0)
-
-        grad_vf = jnp.zeros((ny, nx, 2, 2))
-
-        grad_vf = grad_vf.at[:,:,0,0].set(dudx)
-        grad_vf = grad_vf.at[:,:,0,1].set(dudy)
-        grad_vf = grad_vf.at[:,:,1,0].set(dvdx)
-        grad_vf = grad_vf.at[:,:,1,1].set(dvdy)
-
-        return grad_vf
-
-    return jax.jit(cc_vector_field_gradient)
-
-
-def membrane_strain_rate_function(ny, nx, dy, dx, 
-                                  cc_grad,
-                                  extrapolate_over_cf,
-                                  add_uv_ghost_cells):
-
-    def membrane_sr_tensor(u, v):
-        u = u.reshape((ny, nx))
-        v = v.reshape((ny, nx))
-
-        u = extrapolate_over_cf(u)
-        v = extrapolate_over_cf(v)
-
-        u, v = add_uv_ghost_cells(u, v)
-
-        dudx, dudy = cc_grad(u)
-        dvdx, dvdy = cc_grad(v)
-        
-        
-        msr_tensor = jnp.zeros((ny, nx, 2, 2))
-
-
-        #TODO: CHECK THAT FACTOR OF 4 AND 2! ARE THEY 2 AND 1 REALLY?
-        #I think this is right.
-        msr_tensor = msr_tensor.at[:,:,0,0].set(4*dudx + 2*dvdy)
-        msr_tensor = msr_tensor.at[:,:,0,1].set( dudy + dvdx )
-        msr_tensor = msr_tensor.at[:,:,1,0].set( dudy + dvdx )
-        msr_tensor = msr_tensor.at[:,:,1,1].set(4*dvdy + 2*dudx)
-
-        return msr_tensor
-
-    return jax.jit(membrane_sr_tensor)
-
-
-def divergence_of_tensor_field_function(ny, nx, dy, dx, periodic_x=False):
-    def div_tensor_field(tf):
-        #these have to be 2d scalar fields, of course
-        tf_xx = tf[...,0,0]
-        tf_xy = tf[...,0,1]
-        tf_yx = tf[...,1,0]
-        tf_yy = tf[...,1,1]
-
-        shape_0, shape_1 = tf_xx.shape
-
-        #NOTE: This is done assuming basically everything of interest and its
-        #gradient is zero at the boundaries
-
-        dx_tf_xx = jnp.zeros((shape_0, shape_1))
-        dx_tf_xx = dx_tf_xx.at[:,1:-1].set((tf_xx[:,2:] - tf_xx[:,:-2])/(2*dx))
-    
-        dx_tf_xy = jnp.zeros((shape_0, shape_1))
-        dx_tf_xy = dx_tf_xy.at[:,1:-1].set((tf_xy[:,2:] - tf_xy[:,:-2])/(2*dx))
-        
-        dy_tf_yx = jnp.zeros((shape_0, shape_1))
-        dy_tf_yx = dy_tf_yx.at[1:-1,:].set((tf_yx[:-2,:] - tf_yx[2:,:])/(2*dy))
-        
-        dy_tf_yy = jnp.zeros((shape_0, shape_1))
-        dy_tf_yy = dy_tf_yy.at[1:-1,:].set((tf_yy[:-2,:] - tf_yy[2:,:])/(2*dy))
-        
-        if periodic_x:
-            dx_tf_xx = dx_tf_xx.at[:,0].set((tf_xx[:,1] - tf_xx[:,-1])/(2*dx))
-            dx_tf_xx = dx_tf_xx.at[:,-1].set((tf_xx[:,0] - tf_xx[:,-2])/(2*dx))
-
-        return dx_tf_xx+dy_tf_yx, dx_tf_xy+dy_tf_yy
-    return jax.jit(div_tensor_field)
-
-
-def compute_linear_ssa_residuals_function_fc_visc(ny, nx, dy, dx, \
-                                          h_1d, beta,\
-                                          interp_cc_to_fc,
-                                          ew_gradient,\
-                                          ns_gradient,\
-                                          cc_gradient,\
-                                          add_uv_ghost_cells,\
-                                          add_s_ghost_cells,\
-                                          extrapolate_over_cf):
-
-    def compute_linear_ssa_residuals(u_1d, v_1d, mu_ew, mu_ns, cc_rhs):
-
-        cc_rhs_x = cc_rhs[:nx*ny].reshape((ny, nx))
-        cc_rhs_y = cc_rhs[nx*ny:].reshape((ny, nx))
-
-        u = u_1d.reshape((ny, nx))
-        v = v_1d.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-
-        #volume_term
-        volume_x = - ( beta * u + cc_rhs_x ) * dx * dy
-        volume_y = - ( beta * v + cc_rhs_y ) * dy * dx
-
-        #momentum_term
-        u, v = add_uv_ghost_cells(u, v)
-
-        #get thickness on the faces
-        h = add_s_ghost_cells(h)
-        h_ew, h_ns = interp_cc_to_fc(h)
-        
-        #various face-centred derivatives
-        dudx_ew, dudy_ew = ew_gradient(u)
-        dvdx_ew, dvdy_ew = ew_gradient(v)
-        dudx_ns, dudy_ns = ns_gradient(u)
-        dvdx_ns, dvdy_ns = ns_gradient(v)
-
-        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
-
-        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
-        
-        
-        x_mom_residual = visc_x + volume_x
-        y_mom_residual = visc_y + volume_y
-
-        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1)
-
-    return jax.jit(compute_linear_ssa_residuals)
-
-def compute_linear_ssa_residuals_function_fc_visc_no_rhs(ny, nx, dy, dx, \
-                                          h_1d, beta,\
-                                          interp_cc_to_fc,
-                                          ew_gradient,\
-                                          ns_gradient,\
-                                          cc_gradient,\
-                                          add_uv_ghost_cells,\
-                                          add_s_ghost_cells,\
-                                          extrapolate_over_cf):
-
-    def compute_linear_ssa_residuals(u_1d, v_1d, mu_ew, mu_ns):
-
-        u = u_1d.reshape((ny, nx))
-        v = v_1d.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-
-        
-        s_gnd = h + b #b is globally defined
-        s_flt = h * (1-c.RHO_I/c.RHO_W)
-        s = jnp.maximum(s_gnd, s_flt)
-        
-        s = add_s_ghost_cells(s)
-        #jax.debug.print("s: {x}",x=s)
-
-        dsdx, dsdy = cc_gradient(s)
-        #jax.debug.print("dsdx: {x}",x=dsdx)
-        #sneakily fudge this:
-        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
-        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
-        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
-        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
-        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
-        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
-        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
-        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
-
-
-        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
-        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
-        
-        #momentum_term
-        u, v = add_uv_ghost_cells(u, v)
-
-        #get thickness on the faces
-        h = add_s_ghost_cells(h)
-        h_ew, h_ns = interp_cc_to_fc(h)
-        
-
-        #various face-centred derivatives
-        dudx_ew, dudy_ew = ew_gradient(u)
-        dvdx_ew, dvdy_ew = ew_gradient(v)
-        dudx_ns, dudy_ns = ns_gradient(u)
-        dvdx_ns, dvdy_ns = ns_gradient(v)
-
-        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
-
-        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
-        
-        
-        x_mom_residual = visc_x + volume_x
-        y_mom_residual = visc_y + volume_y
-
-        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1)
-
-    return jax.jit(compute_linear_ssa_residuals)
-
-def compute_u_v_residuals_function_linear_mu(ny, nx, dy, dx, \
-                                   h_1d, beta,\
-                                   interp_cc_to_fc,\
-                                   ew_gradient,\
-                                   ns_gradient,\
-                                   cc_gradient,\
-                                   add_uv_ghost_cells,
-                                   add_s_ghost_cells,
-                                   extrp_over_cf):
-    
-    def compute_u_v_residuals(u_1d, v_1d, mu_ew, mu_ns):
-
-        u = u_1d.reshape((ny, nx))
-        v = v_1d.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-
-
-        s_gnd = h + b #b is globally defined
-        s_flt = h * (1-c.RHO_I/c.RHO_W)
-        s = jnp.maximum(s_gnd, s_flt)
-        
-        s = add_s_ghost_cells(s)
-        #jax.debug.print("s: {x}",x=s)
-
-        dsdx, dsdy = cc_gradient(s)
-        #jax.debug.print("dsdx: {x}",x=dsdx)
-        #sneakily fudge this:
-        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
-        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
-        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
-        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
-        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
-        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
-        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
-        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
-
-        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
-        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
-
-        #obvs not going to do anything in the no-cf case
-        u = extrp_over_cf(u)
-        v = extrp_over_cf(v)
-        #momentum_term
-        u, v = add_uv_ghost_cells(u, v)
-
-        #interpolate things onto face-cenres
-        h = add_s_ghost_cells(h)
-        h_ew, h_ns = interp_cc_to_fc(h)
-        #remove those ghost cells again!
-        h = h[1:-1,1:-1]
-        #jax.debug.print("h_ew = {x}",x=h_ew)
-        #jax.debug.print("h = {x}",x=h)
-
-        #to account for calving front boundary condition, set effective viscosities
-        #of faces of all cells with zero thickness to zero:
-        #Again, shouldn't do owt when there's no calving front
-        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h==0, 0, mu_ew[:, 1:]))
-        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h==0, 0, mu_ew[:,:-1]))
-        mu_ns = mu_ns.at[1:, :].set(jnp.where(h==0, 0, mu_ns[1:, :]))
-        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h==0, 0, mu_ns[:-1,:]))
-        
-        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
-
-        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
-
-        x_mom_residual = visc_x + volume_x
-        y_mom_residual = visc_y + volume_y
-
-
-        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1)
-
-    return jax.jit(compute_u_v_residuals)
-
-def compute_u_v_residuals_function(ny, nx, dy, dx, \
-                                   h_1d, beta,\
-                                   interp_cc_to_fc,\
-                                   ew_gradient,\
-                                   ns_gradient,\
-                                   cc_gradient,\
-                                   add_uv_ghost_cells,
-                                   add_s_ghost_cells,
-                                   extrp_over_cf):
-    
-    def compute_u_v_residuals(u_1d, v_1d, q):
-
-        mucoef = mucoef_0*jnp.exp(q)
-
-        u = u_1d.reshape((ny, nx))
-        v = v_1d.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-
-
-        s_gnd = h + b #b is globally defined
-        s_flt = h * (1-c.RHO_I/c.RHO_W)
-        s = jnp.maximum(s_gnd, s_flt)
-        
-        s = add_s_ghost_cells(s)
-        #jax.debug.print("s: {x}",x=s)
-
-        dsdx, dsdy = cc_gradient(s)
-        #jax.debug.print("dsdx: {x}",x=dsdx)
-        #sneakily fudge this:
-        dsdx = dsdx.at[-1,:].set(dsdx[-2,:])
-        dsdx = dsdx.at[0, :].set(dsdx[1 ,:])
-        dsdx = dsdx.at[:, 0].set(dsdx[:, 1])
-        dsdx = dsdx.at[:,-1].set(dsdx[:,-2])
-        dsdy = dsdy.at[-1,:].set(dsdy[-2,:])
-        dsdy = dsdy.at[0, :].set(dsdy[1 ,:])
-        dsdy = dsdy.at[:, 0].set(dsdy[:, 1])
-        dsdy = dsdy.at[:,-1].set(dsdy[:,-2])
-
-        volume_x = - (beta * u + c.RHO_I * c.g * h * dsdx) * dx * dy
-        volume_y = - (beta * v + c.RHO_I * c.g * h * dsdy) * dy * dx
-
-        #obvs not going to do anything in the no-cf case
-        u = extrp_over_cf(u)
-        #jax.debug.print("{x}",x=u[:,-3:])
-
-        v = extrp_over_cf(v)
-        #momentum_term
-        u, v = add_uv_ghost_cells(u, v)
-
-        #various face-centred derivatives
-        dudx_ew, dudy_ew = ew_gradient(u)
-        dvdx_ew, dvdy_ew = ew_gradient(v)
-        dudx_ns, dudy_ns = ns_gradient(u)
-        dvdx_ns, dvdy_ns = ns_gradient(v)
-
-        #jax.debug.print("dudxew = {x}",x=dudx_ew)
-
-        #interpolate things onto face-cenres
-        h = add_s_ghost_cells(h)
-        h_ew, h_ns = interp_cc_to_fc(h)
-        #remove those ghost cells again!
-        h = h[1:-1,1:-1]
-        #jax.debug.print("h_ew = {x}",x=h_ew)
-        #jax.debug.print("h = {x}",x=h)
-
-        mucoef = add_s_ghost_cells(mucoef)
-        mucoef_ew, mucoef_ns = interp_cc_to_fc(mucoef)
-        #jax.debug.print("mucoef_ew = {x}",x=mucoef_ew)
-
-        #calculate face-centred viscosity:
-        mu_ew = B * mucoef_ew * (dudx_ew**2 + dvdy_ew**2 + dudx_ew*dvdy_ew +\
-                    0.25*(dudy_ew+dvdx_ew)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        mu_ns = B * mucoef_ns * (dudx_ns**2 + dvdy_ns**2 + dudx_ns*dvdy_ns +\
-                    0.25*(dudy_ns+dvdx_ns)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        
-        #to account for calving front boundary condition, set effective viscosities
-        #of faces of all cells with zero thickness to zero:
-        #Again, shouldn't do owt when there's no calving front
-        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h==0, 0, mu_ew[:, 1:]))
-        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h==0, 0, mu_ew[:,:-1]))
-        mu_ns = mu_ns.at[1:, :].set(jnp.where(h==0, 0, mu_ns[1:, :]))
-        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h==0, 0, mu_ns[:-1,:]))
-        
-
-        visc_x = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(2*dudx_ew[:, 1:] + dvdy_ew[:, 1:])*dy   -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(2*dudx_ew[:,:-1] + dvdy_ew[:,:-1])*dy   +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(dudy_ns[:-1,:] + dvdx_ns[:-1,:])*0.5*dx -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(dudy_ns[1:, :] + dvdx_ns[1:, :])*0.5*dx
-
-        visc_y = 2 * mu_ew[:, 1:]*h_ew[:, 1:]*(dudy_ew[:, 1:] + dvdx_ew[:, 1:])*0.5*dy -\
-                 2 * mu_ew[:,:-1]*h_ew[:,:-1]*(dudy_ew[:,:-1] + dvdx_ew[:,:-1])*0.5*dy +\
-                 2 * mu_ns[:-1,:]*h_ns[:-1,:]*(2*dvdy_ns[:-1,:] + dudx_ns[:-1,:])*dx   -\
-                 2 * mu_ns[1:, :]*h_ns[1:, :]*(2*dvdy_ns[1:, :] + dudx_ns[1:, :])*dx
-
-        x_mom_residual = visc_x + volume_x
-        y_mom_residual = visc_y + volume_y
-
-
-        return x_mom_residual.reshape(-1), y_mom_residual.reshape(-1)
-
-    return jax.jit(compute_u_v_residuals)
-
-
-def cc_viscosity_function(ny, nx, dy, dx, cc_vector_field_gradient):
-    def cc_viscosity(q, u, v):
-
-        vfg = cc_vector_field_gradient(u, v)
-        
-        mu = B * mucoef_0 * jnp.exp(q) * (vfg[:,:,0,0]**2 + vfg[:,:,1,1]**2 + vfg[:,:,0,0]*vfg[:,:,1,1] + \
-                           0.25*(vfg[:,:,0,1] + vfg[:,:,1,0])**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-
-        #NOTE: the effective viscosity isn't actually set to zero here in the ice-free regions,
-        #but wherever it's used, it should be multiplied by a zero thickness there...
-
-        return mu
-    return jax.jit(cc_viscosity)
-
-
-def fc_viscosity_function(ny, nx, dy, dx, extrp_over_cf, add_uv_ghost_cells,
-                          add_mucoef_ghost_cells,
-                          interp_cc_to_fc, ew_gradient, ns_gradient, h_1d):
-    def fc_viscosity(q, u, v):
-        mucoef = mucoef_0*jnp.exp(q)
-        mucoef = add_mucoef_ghost_cells(mucoef)
-        mucoef_ew, mucoef_ns = interp_cc_to_fc(mucoef)
-        
-        u = u.reshape((ny, nx))
-        v = v.reshape((ny, nx))
-        h = h_1d.reshape((ny, nx))
-
-        u = extrp_over_cf(u)
-        v = extrp_over_cf(v)
-        #and add the ghost cells in
-        u, v = add_uv_ghost_cells(u, v)
-
-        #various face-centred derivatives
-        dudx_ew, dudy_ew = ew_gradient(u)
-        dvdx_ew, dvdy_ew = ew_gradient(v)
-        dudx_ns, dudy_ns = ns_gradient(u)
-        dvdx_ns, dvdy_ns = ns_gradient(v)
-        
-        #calculate face-centred viscosity:
-        mu_ew = B * mucoef_ew * (dudx_ew**2 + dvdy_ew**2 + dudx_ew*dvdy_ew +\
-                    0.25*(dudy_ew+dvdx_ew)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-        mu_ns = B * mucoef_ns * (dudx_ns**2 + dvdy_ns**2 + dudx_ns*dvdy_ns +\
-                    0.25*(dudy_ns+dvdx_ns)**2 + c.EPSILON_VISC**2)**(0.5*(1/nvisc - 1))
-
-        #to account for calving front boundary condition, set effective viscosities
-        #of faces of all cells with zero thickness to zero:
-        mu_ew = mu_ew.at[:, 1:].set(jnp.where(h==0, 0, mu_ew[:, 1:]))
-        mu_ew = mu_ew.at[:,:-1].set(jnp.where(h==0, 0, mu_ew[:,:-1]))
-        mu_ns = mu_ns.at[1:, :].set(jnp.where(h==0, 0, mu_ns[1:, :]))
-        mu_ns = mu_ns.at[:-1,:].set(jnp.where(h==0, 0, mu_ns[:-1,:]))
-
-        return mu_ew, mu_ns
-    return jax.jit(fc_viscosity)
-
-
-def print_residual_things(residual, rhs, init_residual, i):
-    old_residual = residual
-    residual = jnp.max(jnp.abs(-rhs))
-
-    if i==0:
-        init_residual = residual.copy()
-        print("Initial residual: {}".format(residual))
-    else:
-        print("residual: {}".format(residual))
-        print("Residual reduction factor: {}".format(old_residual/residual))
-    print("------")
-
-    return old_residual, residual, init_residual
-
-
-
-
-def make_newton_velocity_solver_function_custom_vjp(ny, nx, dy, dx,\
-                                                    h, C,\
-                                                    n_iterations):
-
-    beta_eff = C.copy()
-    h_1d = h.reshape(-1)
-
-    #functions for various things:
-    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
-    cc_gradient                                = cc_gradient_function(dy, dx)
-    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
-    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny, nx)
-    #add_uv_ghost_cells                         = apply_scalar_ghost_cells_to_vector(
-    #                                                add_ghost_cells_periodic_dirichlet_function(ny,nx)
-    #                                             )
-    extrapolate_over_cf                        = extrapolate_over_cf_function(h)
-
-    get_u_v_residuals = compute_u_v_residuals_function(ny, nx, dy, dx,\
-                                                       h_1d, beta_eff,\
-                                                       interp_cc_to_fc,\
-                                                       ew_gradient, ns_gradient,\
-                                                       cc_gradient,\
-                                                       add_uv_ghost_cells,\
-                                                       add_scalar_ghost_cells,\
-                                                       extrapolate_over_cf)
-    
-
-    #############
-    #setting up bvs and coords for a single block of the jacobian
-    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
-                                                                                  periodic_x=False)
-
-    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
-    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
-    mask = (i_coordinate_sets>=0)
-
-
-    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
-                                                        basis_vectors,\
-                                                        i_coordinate_sets,\
-                                                        j_coordinate_sets,\
-                                                        mask,\
-                                                        2,
-                                                        active_indices=(0,1)
-                                                       )
-    #sparse_jacrev = jax.jit(sparse_jacrev)
-
-
-    i_coordinate_sets = i_coordinate_sets[mask]
-    j_coordinate_sets = j_coordinate_sets[mask]
-    #############
-
-    coords = jnp.stack([
-                    jnp.concatenate(
-                                [i_coordinate_sets,         i_coordinate_sets,\
-                                 i_coordinate_sets+(ny*nx), i_coordinate_sets+(ny*nx)]
-                                   ),\
-                    jnp.concatenate(
-                                [j_coordinate_sets, j_coordinate_sets+(ny*nx),\
-                                 j_coordinate_sets, j_coordinate_sets+(ny*nx)]
-                                   )
-                       ])
-
-   
-    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*2, ny*nx*2),\
-                                                              ksp_type="bcgs",\
-                                                              preconditioner="hypre",\
-                                                              precondition_only=False)
-
-
-
-    @custom_vjp
-    def solver(q, u_trial, v_trial):
-        u_trial = jnp.where(h>1e-10, u_trial, 0)
-        v_trial = jnp.where(h>1e-10, v_trial, 0)
-
-        u_1d = u_trial.copy().reshape(-1)
-        v_1d = v_trial.copy().reshape(-1)
-
-        residual = jnp.inf
-        init_res = 0
-
-        for i in range(n_iterations):
-
-            dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
-                                                           (u_1d, v_1d, q)
-                                                          )
-
-            nz_jac_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-                                             dJv_du[mask], dJv_dv[mask]])
-
-
-            rhs = -jnp.concatenate(get_u_v_residuals(u_1d, v_1d, q))
-
-            #print(jnp.max(rhs))
-            #raise
-
-            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
-
-
-            du = la_solver(nz_jac_values, rhs)
-
-            u_1d = u_1d+du[:(ny*nx)]
-            v_1d = v_1d+du[(ny*nx):]
-
-
-        res_final = jnp.max(jnp.abs(jnp.concatenate(
-                                    get_u_v_residuals(u_1d, v_1d, q)
-                                                   )
-                                   )
-                           )
-        print("----------")
-        print("Final residual: {}".format(res_final))
-        print("Total residual reduction factor: {}".format(init_res/res_final))
-        print("----------")
-        
-        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx))
-
-
-
-    def solver_fwd(q, u_trial, v_trial):
-        u_trial = jnp.where(h>1e-10, u_trial, 0)
-        v_trial = jnp.where(h>1e-10, v_trial, 0)
-        
-        u, v = solver(q, u_trial, v_trial)
-
-        dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
-                                                       (u.reshape(-1), v.reshape(-1), q)
-                                                      )
-        dJ_dvel_nz_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-                                             dJv_du[mask], dJv_dv[mask]])
-
-
-        fwd_residuals = (u, v, dJ_dvel_nz_values, q)
-        #fwd_residuals = (u, v, q)
-
-        return (u, v), fwd_residuals
-
-
-    def solver_bwd(res, cotangent):
-        
-        u, v, dJ_dvel_nz_values, q = res
-        #u, v, q = res
-
-        u_bar, v_bar = cotangent
-        
-
-        #dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(get_u_v_residuals, \
-        #                                               (u.reshape(-1), v.reshape(-1), q)
-        #                                              )
-        #dJ_dvel_nz_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-        #                                     dJv_du[mask], dJv_dv[mask]])
-
-
-        lambda_ = la_solver(dJ_dvel_nz_values,
-                            -jnp.concatenate([u_bar, v_bar]),
-                            transpose=True)
-
-        lambda_u = lambda_[:(ny*nx)]
-        lambda_v = lambda_[(ny*nx):]
-
-
-        #phi_bar = (dG/dphi)^T lambda
-        _, pullback_function = jax.vjp(get_u_v_residuals,
-                                         u.reshape(-1), v.reshape(-1), q
-                                      )
-        _, _, mu_bar = pullback_function((lambda_u, lambda_v))
-        
-#        #bwd has to return a tuple of cotangents for each primal input
-#        #of solver, so have to return this 1-tuple:
-#        return (mu_bar.reshape((ny, nx)), )
-
-        #I wonder if I can get away with just returning None for u_trial_bar and v_trial_bar...
-        return (mu_bar.reshape((ny, nx)), None, None)
-
-
-    solver.defvjp(solver_fwd, solver_bwd)
-
-    return solver
-
-
-
-def generic_newton_solver_no_cjvp(ny, nx, sparse_jacrev, mask, la_solver):
-
-    def solver(u_trial, v_trial, residuals_function, n_iterations, residuals_fct_args):
-
-        u_1d = u_trial.copy().reshape(-1)
-        v_1d = v_trial.copy().reshape(-1)
-        
-        residual = jnp.inf
-        init_res = 0
-        
-        for i in range(n_iterations):
-            dJu_du, dJv_du, dJu_dv, dJv_dv = sparse_jacrev(residuals_function,
-                                                           (u_1d, v_1d,
-                                                            *residuals_fct_args)
-                                                          )
-
-            nz_jac_values = jnp.concatenate([dJu_du[mask], dJu_dv[mask],\
-                                             dJv_du[mask], dJv_dv[mask]])
-
-
-
-
-            rhs = -jnp.concatenate(residuals_function(u_1d, v_1d,
-                                                      *residuals_fct_args))
-
-
-            old_residual, residual, init_res = print_residual_things(residual, rhs, init_res, i)
-
-            du = la_solver(nz_jac_values, rhs)
-
-            u_1d = u_1d+du[:(ny*nx)]
-            v_1d = v_1d+du[(ny*nx):]
-
-
-        res_final = jnp.max(jnp.abs(jnp.concatenate(
-                                    residuals_function(u_1d, v_1d, *residuals_fct_args)
-                                                   )
-                                   )
-                           )
-        print("----------")
-        print("Final residual: {}".format(res_final))
-        print("Total residual reduction factor: {}".format(init_res/res_final))
-        print("----------")
-
-        return u_1d.reshape((ny, nx)), v_1d.reshape((ny, nx))
-
-    return solver
-
-
-def forward_adjoint_and_second_order_adjoint_solvers(ny, nx, dy, dx, h, C, n_iterations):
-
-    beta_eff = C.copy()
-    h_1d = h.reshape(-1)
-    
-    #functions for various things:
-    interp_cc_to_fc                            = interp_cc_with_ghosts_to_fc_function(ny, nx)
-    ew_gradient, ns_gradient                   = fc_gradient_functions(dy, dx)
-    cc_gradient                                = cc_gradient_function(dy, dx)
-    add_uv_ghost_cells, add_cont_ghost_cells   = add_ghost_cells_fcts(ny, nx)
-    add_scalar_ghost_cells                     = add_ghost_cells_periodic_continuation_function(ny,nx)
-    #add_uv_ghost_cells                         = apply_scalar_ghost_cells_to_vector(
-    #                                                add_ghost_cells_periodic_dirichlet_function(ny,nx)
-    #                                             )
-    extrapolate_over_cf                        = extrapolate_over_cf_function(h)
-    cc_vector_field_gradient                   = cc_vector_field_gradient_function(ny, nx, dy,
-                                                                                   dx, cc_gradient, 
-                                                                                   extrapolate_over_cf,
-                                                                                   add_uv_ghost_cells)
-    membrane_strain_rate                       = membrane_strain_rate_function(ny, nx, dy, dx,
-                                                                               cc_gradient,
-                                                                               extrapolate_over_cf,
-                                                                               add_uv_ghost_cells)
-    div_tensor_field                           = divergence_of_tensor_field_function(ny, nx, dy, dx,
-                                                                                     periodic_x=False)
-
-    #calculate cell-centred viscosity based on velocity and q
-    cc_viscosity = cc_viscosity_function(ny, nx, dy, dx, cc_vector_field_gradient)
-    fc_viscosity = fc_viscosity_function(ny, nx, dy, dx, extrapolate_over_cf, add_uv_ghost_cells,
-                                         add_scalar_ghost_cells,
-                                         interp_cc_to_fc, ew_gradient, ns_gradient, h_1d)
-
-    get_u_v_residuals = compute_u_v_residuals_function(ny, nx, dy, dx,\
-                                                       h_1d, beta_eff,\
-                                                       interp_cc_to_fc,\
-                                                       ew_gradient, ns_gradient,\
-                                                       cc_gradient,\
-                                                       add_uv_ghost_cells,\
-                                                       add_scalar_ghost_cells,\
-                                                       extrapolate_over_cf)
-    
-    linear_ssa_residuals = compute_linear_ssa_residuals_function_fc_visc(ny, nx, dy, dx,\
-                                                       h_1d, beta_eff,\
-                                                       interp_cc_to_fc,\
-                                                       ew_gradient, ns_gradient,\
-                                                       cc_gradient,\
-                                                       add_uv_ghost_cells,\
-                                                       add_scalar_ghost_cells,\
-                                                       extrapolate_over_cf)
-
-    linear_ssa_residuals_no_rhs = compute_linear_ssa_residuals_function_fc_visc_no_rhs(
-                                                       ny, nx, dy, dx,\
-                                                       h_1d, beta_eff,\
-                                                       interp_cc_to_fc,\
-                                                       ew_gradient, ns_gradient,\
-                                                       cc_gradient,\
-                                                       add_uv_ghost_cells,\
-                                                       add_scalar_ghost_cells,\
-                                                       extrapolate_over_cf
-                                                                                      )
-
-    #############
-    #setting up bvs and coords for a single block of the jacobian
-    basis_vectors, i_coordinate_sets = basis_vectors_and_coords_2d_square_stencil(ny, nx, 1,
-                                                                                  periodic_x=False)
-
-    i_coordinate_sets = jnp.concatenate(i_coordinate_sets)
-    j_coordinate_sets = jnp.tile(jnp.arange(ny*nx), len(basis_vectors))
-    mask = (i_coordinate_sets>=0)
-
-
-    sparse_jacrev = make_sparse_jacrev_fct_shared_basis(
-                                                        basis_vectors,\
-                                                        i_coordinate_sets,\
-                                                        j_coordinate_sets,\
-                                                        mask,\
-                                                        2,
-                                                        active_indices=(0,1)
-                                                       )
-    #sparse_jacrev = jax.jit(sparse_jacrev)
-
-
-    i_coordinate_sets = i_coordinate_sets[mask]
-    j_coordinate_sets = j_coordinate_sets[mask]
-    #############
-
-    coords = jnp.stack([
-                    jnp.concatenate(
-                                [i_coordinate_sets,         i_coordinate_sets,\
-                                 i_coordinate_sets+(ny*nx), i_coordinate_sets+(ny*nx)]
-                                   ),\
-                    jnp.concatenate(
-                                [j_coordinate_sets, j_coordinate_sets+(ny*nx),\
-                                 j_coordinate_sets, j_coordinate_sets+(ny*nx)]
-                                   )
-                       ])
-
-   
-
-
-    la_solver = create_sparse_petsc_la_solver_with_custom_vjp(coords, (ny*nx*2, ny*nx*2),\
-                                                              ksp_type="bcgs",\
-                                                              preconditioner="hypre",\
-                                                              precondition_only=False,
-                                                              monitor_ksp=False)
-
-
-    newton_solver = generic_newton_solver_no_cjvp(ny, nx, sparse_jacrev, mask, la_solver)
-
-    #picard_solver = make_picard_solver(ny, nx, sparse_jacrev, mask, la_solver, get_u_v_residuals, fc_visc)
-
-    def solve_fwd_problem(q, u_trial, v_trial):
-        u_trial = jnp.where(h>1e-10, u_trial, 0)
-        v_trial = jnp.where(h>1e-10, v_trial, 0)
-        
-        #u, v = newton_solver(u_trial, v_trial, get_u_v_residuals, n_iterations, (q,), coords)
-        u, v = newton_solver(u_trial, v_trial, get_u_v_residuals, n_iterations, (q,))
-        return u.reshape((ny,nx)), v.reshape((ny,nx))
-
-
-    def solve_fwd_problem_picard(q, u, v, n_pic_its=16):
-        mu_bar_ew = jnp.zeros((ny,nx+1))+3e5
-        mu_bar_ns = jnp.zeros((ny+1,nx))+3e5
-        for i in range(n_pic_its):
-            #calculate viscosity
-
-            #plt.imshow(mu_bar_ns)
-            #plt.colorbar()
-            #plt.show()
-
-            #solve adjoint problem
-            u_new, v_new = newton_solver(u.reshape(-1), v.reshape(-1),
-                                 linear_ssa_residuals_no_rhs, 1, (mu_bar_ew, mu_bar_ns), coords)
-
-            #u = 0.9*u + 0.1*u_new
-            #v = 0.9*v + 0.1*v_new
-            u = u_new
-            v = v_new
-            
-            mu_bar_ew, mu_bar_ns = fc_viscosity(q, u, v)
-
-        return u.reshape((ny,nx)), v.reshape((ny,nx))
-
-
-    def solve_adjoint_problem(q, u, v, lx_trial, ly_trial,
-                              functional:callable, additional_fctl_args=None):
-        #calculate viscosity
-        mu_bar_ew, mu_bar_ns = fc_viscosity(q, u, v)
-
-        #right-hand-side (\partial_u J)
-        if additional_fctl_args is None:
-            argz = (u.reshape(-1), v.reshape(-1), q,)
-        else:
-            argz = (u.reshape(-1), v.reshape(-1), q, *additional_fctl_args)
-
-        dJdu, dJdv = jax.grad(functional, argnums=(0,1))(*argz)
-
-        
-        rhs = - jnp.concatenate([dJdu, dJdv])
-
-
-        #solve adjoint problem
-        lx, ly = newton_solver(lx_trial.reshape(-1), ly_trial.reshape(-1),
-                               linear_ssa_residuals, 1, (mu_bar_ew, mu_bar_ns, rhs))
-
-        mu_bar = cc_viscosity(q, u, v)
-
-        #plt.imshow(mu_bar)
-        #plt.colorbar()
-        #plt.show()
-
-        #plt.imshow(mu_bar_ns)
-        #plt.colorbar()
-        #plt.show()
-
-
-        #calculate gradient
-        dJdq = jax.grad(functional, argnums=2)(*argz) \
-               - mu_bar * h * double_dot_contraction(cc_vector_field_gradient(lx.reshape(-1), ly.reshape(-1)),
-                                                           membrane_strain_rate(u.reshape(-1), v.reshape(-1))
-                                                    )
-
-        return lx.reshape((ny,nx)), ly.reshape((ny,nx)), dJdq
-
-
-    def solve_soa_problem(q, u, v, lx, ly, perturbation_direction,
-                          functional:callable, 
-                          additional_fctl_args=None):
-        #calculate viscosity
-        mu_bar = cc_viscosity(q, u, v)
-        mu_bar_ew, mu_bar_ns = fc_viscosity(q, u, v)
-        
-        #right-hand-side (\partial_u J)
-        if additional_fctl_args is None:
-            argz = (u.reshape(-1), v.reshape(-1), q)
-        else:
-            argz = (u.reshape(-1), v.reshape(-1), q, *additional_fctl_args)
-
-        
-        #solve first equation for mu
-        rhs_x, rhs_y = div_tensor_field((h * mu_bar * perturbation_direction)[...,None,None] *\
-                                        membrane_strain_rate(u.reshape(-1), v.reshape(-1)))
-        rhs = - jnp.concatenate([rhs_x.reshape(-1), rhs_y.reshape(-1)])
-
-        x_trial = jnp.zeros_like(mu_bar).reshape(-1)
-        y_trial = jnp.zeros_like(x_trial)
-
-
-        mu_x, mu_y = newton_solver(x_trial, y_trial, linear_ssa_residuals,
-                                   1, (mu_bar_ew, mu_bar_ns, rhs))
-
-
-        #solve second equation for beta
-        #NOTE: make functional essentially a function just of u,v to avoid the
-        #difficulties with what to do with q gradients
-        functional_fixed_q = lambda u, v: functional(u, v, q)
-        gradient_j = jax.grad(functional_fixed_q, argnums=(0,1))
-        direct_hvp_x, direct_hvp_y = jax.jvp(gradient_j,
-                                             (u.reshape(-1), v.reshape(-1)),
-                                             (mu_x.reshape(-1), mu_y.reshape(-1)))[1]
-        rhs_1_x, rhs_1_y = div_tensor_field((h * mu_bar * perturbation_direction)[...,None,None] *\
-                                        membrane_strain_rate(lx.reshape(-1), ly.reshape(-1))
-                                           )
-
-        rhs = - jnp.concatenate([(rhs_1_x.reshape(-1) + direct_hvp_x),
-                                 (rhs_1_y.reshape(-1) + direct_hvp_y)])
-
-        beta_x, beta_y = newton_solver(x_trial, y_trial, linear_ssa_residuals,
-                                       1, (mu_bar_ew, mu_bar_ns, rhs))
-
-
-        #calculate hessian-vector-product
-        functional_fixed_vel = lambda q: functional(u.reshape(-1), v.reshape(-1), q)
-        direct_hvp_part = jax.jvp(jax.grad(functional_fixed_vel, argnums=0),
-                                  (q,),
-                                  (perturbation_direction,))[1]
-        #NOTE: the first term in the brackets is zero if we're thinking about q rather than q
-        hvp = direct_hvp_part - mu_bar * h * (\
-              perturbation_direction * double_dot_contraction(
-                                              cc_vector_field_gradient(lx.reshape(-1), ly.reshape(-1)),
-                                              membrane_strain_rate(u.reshape(-1), v.reshape(-1))
-                                                             ) +\
-              double_dot_contraction(
-                            cc_vector_field_gradient(lx.reshape(-1), ly.reshape(-1)),
-                            membrane_strain_rate(mu_x.reshape(-1), mu_y.reshape(-1))
-                                    ) +\
-              double_dot_contraction(
-                            cc_vector_field_gradient(beta_x.reshape(-1), beta_y.reshape(-1)),
-                            membrane_strain_rate(u.reshape(-1), v.reshape(-1))
-                                    )
-                                              )
-
-        return hvp
-
-    return solve_fwd_problem, solve_adjoint_problem, solve_soa_problem
-    #return solve_fwd_problem_picard, solve_adjoint_problem, solve_soa_problem
-
-
-
-
-
-
-##NOTE: make everything linear by changing to 1
-nvisc = c.GLEN_N
-#nvisc = 1.001
-
-A = c.A_COLD
-B = 0.5 * (A**(-1/nvisc))
-
 
 def ice_shelf():
     lx = 150_000
     ly = 200_000
     
-    resolution = 4000 #m
+    resolution = 4_000 #m
     
     nr = int(ly/resolution)
     nc = int(lx/resolution)
@@ -1514,16 +69,16 @@ def ice_shelf():
     thk_profile = 500# - 300*x/lx
     thk = jnp.zeros((nr, nc))+thk_profile
     thk = thk.at[:, -2:].set(0)
-    
+   
     b = jnp.zeros_like(thk)-600
     
     mucoef = jnp.ones_like(thk)
     
     C = jnp.zeros_like(thk)
-    C = C.at[:4, :].set(1e16)
-    C = C.at[:, :4].set(1e16)
-    C = C.at[-4:,:].set(1e16)
-    C = jnp.where(thk==0, 1, C)
+    C = C.at[:4, :].set(1e8)
+    C = C.at[:, :4].set(1e8)
+    C = C.at[-4:,:].set(1e8)
+    C = jnp.where(thk==0, 1e8, C)
 
     #mucoef_profile = 0.5+b_profile.copy()/2000
     mucoef_profile = 1
@@ -1543,9 +98,9 @@ v_init = jnp.zeros_like(b)
 n_iterations = 10
 
 mask = jnp.where(thk>0,1,0)
-mask = binary_erosion(mask)
-mask = binary_erosion(mask)
-mask = binary_erosion(mask)
+#mask = binary_erosion(mask)
+#mask = binary_erosion(mask)
+#mask = binary_erosion(mask)
 mask = mask.astype(int)
 #print(mask)
 #raise
@@ -1559,7 +114,7 @@ def functional(v_field_x, v_field_y, q):
 
 def calculate_gradient_via_foa():
     fwd_solver, adjoint_solver, soa_solver = forward_adjoint_and_second_order_adjoint_solvers(
-                                             nr, nc, delta_y, delta_x, thk, C, n_iterations
+                                      nr, nc, delta_y, delta_x, thk, b, C, n_iterations, mucoef_0
                                                                                              )
     
     print("solving fwd problem:")
@@ -1579,8 +134,8 @@ def calculate_gradient_via_ad():
     solver = make_newton_velocity_solver_function_custom_vjp(nr, nc,
                                                              delta_y,
                                                              delta_x,
-                                                             thk, C,
-                                                             n_iterations)
+                                                             thk, b, C,
+                                                             n_iterations, mucoef_0)
 
     def reduced_functional(q):
         u_out, v_out = solver(q, u_init, v_init)
@@ -1603,7 +158,7 @@ def calculate_gradient_via_ad():
 
 def calculate_hvp_via_soa():
     fwd_solver, adjoint_solver, soa_solver = forward_adjoint_and_second_order_adjoint_solvers(
-                                             nr, nc, delta_y, delta_x, thk, C, n_iterations
+                                      nr, nc, delta_y, delta_x, thk, b, C, n_iterations, mucoef_0
                                                                                              )
     
     print("solving fwd problem:")
@@ -1626,7 +181,7 @@ def calculate_hvp_via_soa():
 
     print("solving second-order adjoint problem:")
     pert_dir = gradient.copy()/(jnp.linalg.norm(gradient)*10)
-    pert_dir = pert_dir.at[:,-4:].set(0)
+    #pert_dir = pert_dir.at[:,-4:].set(0)
     hvp = soa_solver(q, u_out, v_out, lx, ly, pert_dir, functional)
     
     #plt.imshow(hvp[:,:], vmin=-3, vmax=3, cmap="twilight_shifted")
@@ -1642,11 +197,11 @@ def calculate_hvp_via_ad():
     solver = make_newton_velocity_solver_function_custom_vjp(nr, nc,
                                                              delta_y,
                                                              delta_x,
-                                                             thk, C,
-                                                             n_iterations)
+                                                             thk, b, C,
+                                                             n_iterations, mucoef_0)
 
     u_out, v_out = solver(q, u_init, v_init)
-    #show_vel_field(u_out, v_out)
+    show_vel_field(u_out, v_out)
     
     def reduced_functional(q):
         u_out, v_out = solver(q, u_init, v_init)
@@ -1668,7 +223,7 @@ def calculate_hvp_via_ad():
 
     
     pert_dir = gradient.copy() / (jnp.linalg.norm(gradient)*10)
-    pert_dir = pert_dir.at[:,-4:].set(0)
+    #pert_dir = pert_dir.at[:,-4:].set(0)
 
     ##finite diff hvp for comparison
     ##plt.imshow(p)
@@ -1756,21 +311,50 @@ calculate_hvp_via_soa()
 raise
 
 
+solver = make_newton_velocity_solver_function_custom_vjp(nr, nc,
+                                                         delta_y,
+                                                         delta_x,
+                                                         thk, b, C,
+                                                         n_iterations, mucoef_0)
+
+#u_out, v_out = solver(q, u_init, v_init)
+#show_vel_field(u_out, v_out)
+##raise
+#jnp.save("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/u_big_new.npy", u_out)
+#jnp.save("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/v_big_new.npy", v_out)
+#raise
+
+#u_data = jnp.load("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/u_big.npy")
+#v_data = jnp.load("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/v_big.npy")
+
+u_data = jnp.load("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/u_big_new.npy")
+v_data = jnp.load("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/v_big_new.npy")
+
+#show_vel_field(u_data, v_data)
+
+def misfit_functional(v_field_x, v_field_y, q):
+
+    return jnp.sum(mask.reshape(-1) * \
+                   jnp.sqrt((v_field_x.reshape(-1)-u_data.reshape(-1))**2 +\
+                            (v_field_y.reshape(-1)-v_data.reshape(-1))**2 +\
+                            1e-10)
+                  )
+
+
+
 
 def make_hvp_ad_fct():
     solver = make_newton_velocity_solver_function_custom_vjp(nr, nc,
                                                              delta_y,
                                                              delta_x,
-                                                             thk, C,
-                                                             n_iterations)
+                                                             thk, b, C,
+                                                             n_iterations, mucoef_0)
     def reduced_functional(q):
         u_out, v_out = solver(q, u_init, v_init)
-        return functional(u_out, v_out, q)
+        return misfit_functional(u_out, v_out, q)
+        #return functional(u_out, v_out, q)
 
     get_grad = jax.grad(reduced_functional)
-
-    def hess_vec_product(q, perturbation):
-        return jax.grad(lambda m: jnp.vdot(get_grad(m), perturbation))(q)
 
     _, vjp_grad = jax.vjp(get_grad, q)
     
@@ -1783,7 +367,7 @@ def make_hvp_ad_fct():
 
 def make_hvp_soa_fct():
     fwd_solver, adjoint_solver, soa_solver = forward_adjoint_and_second_order_adjoint_solvers(
-                                             nr, nc, delta_y, delta_x, thk, C, n_iterations
+                                      nr, nc, delta_y, delta_x, thk, b, C, n_iterations, mucoef_0
                                                                                              )
     
     print("solving fwd problem:")
@@ -1795,11 +379,11 @@ def make_hvp_soa_fct():
     lx, ly, gradient = adjoint_solver(q, u_out, v_out,
                                       jnp.zeros_like(u_out),
                                       jnp.zeros_like(u_out),
-                                      functional)
+                                      misfit_functional)
     
     
     def hessian_vector_product(pert_dir):
-        hvp = soa_solver(q, u_out, v_out, lx, ly, pert_dir.reshape((nr,nc)), functional)
+        hvp = soa_solver(q, u_out, v_out, lx, ly, pert_dir.reshape((nr,nc)), misfit_functional)
         return hvp
 
     return hessian_vector_product
@@ -1807,7 +391,7 @@ def make_hvp_soa_fct():
 
 #function for computing hvp from perturbation direction
 #hessian_vector_product = make_hvp_ad_fct()
-#hessian_vector_product_soa = make_hvp_soa_fct()
+hessian_vector_product = make_hvp_soa_fct()
 
 
 
@@ -1823,13 +407,9 @@ dtype = np.float64
 #HVP_jit = jax.jit(lambda x: HVP(x.reshape(nr, nc)).reshape(-1))
 #_ = HVP_jit(np.zeros(n))   # warmup
 
-#H = LinearOperator(
-#    shape=(n, n), dtype=dtype,
-#    matvec=lambda x: np.array(hessian_vector_product(x))  # ensure ndarray to avoid device transfers
-#)
 H = LinearOperator(
     shape=(n, n), dtype=dtype,
-    matvec=lambda x: np.array(hessian_vector_product_soa(x))  # ensure ndarray to avoid device transfers
+    matvec=lambda x: np.array(hessian_vector_product(x))  # ensure ndarray to avoid device transfers
 )
 
 # Largest algebraic eigenvalues (most positive)
@@ -1845,7 +425,7 @@ for i in range(k):
     plt.imshow(eigvecs[...,i])
     plt.colorbar()
     plt.title("lambda={}".format(eigvals[i]))
-    plt.savefig("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/ts_evec_soa_{}.png".format(i))
+    plt.savefig("/users/eetss/new_model_code/src/nm/bits_of_data/hessian_evecs_etc/shelf/ts_evec_soa_big_new_new_{}.png".format(i))
     plt.close()
 
 plt.plot(eigvals[::-1])
